### 2022-05-04 #4371 @editorialbot
[REVIEW]: anndata: Access and store annotated data matrices
**Submitting author:** @falexwolf (<a href="http://orcid.org/0000-0002-8760-7838">F. Alexander Wolf</a>)
**Repository:** https://github.com/scverse/anndata
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.10.9
