### 2022-05-19 60a4198
Restrict bugged pip release
