+++ anndata/_core/anndata.py
@@ -834,7 +834,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
                     Inferred to be: {infer_dtype(value)}
                 """
-                ), # noqa
+            ),  # noqa
                 stacklevel=2,
             )
         # fmt: on
@@ -1337,7 +1337,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             Key for `.obsm`.
         """
         if layer is not None and obsm is not None:
-            raise ValueError('Can not specify both layer and obsm at once.')
+            raise ValueError("Can not specify both layer and obsm at once.")
         if layer is not None:
             X = self.layers[layer]
         elif obsm is not None:
@@ -1348,8 +1348,9 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             X = self.X
         if issparse(X):
             X = X.toarray()
-        return pd.DataFrame(X, index=self.obs_names,
-                            columns=self.var_names if not obsm else None)
+        return pd.DataFrame(
+            X, index=self.obs_names, columns=self.var_names if not obsm else None
+        )
 
     def _get_X(self, use_raw=False, layer=None):
         """\
@@ -1808,13 +1809,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             partial(merge_outer, batch_keys=batch_categories, merge=merge_same),
         )
         out.var = out.var.iloc[
-            :,
-            (
-                out.var.columns.str.extract(pat, expand=False)
-                .fillna("")
-                .argsort(kind="stable")
-            ),
-        ]
+                  :,
+                  (
+                      out.var.columns.str.extract(pat, expand=False)
+                          .fillna("")
+                          .argsort(kind="stable")
+                  ),
+                  ]
 
         return out
 
