### 2022-05-31 scverse/anndata#769 @rcannood
Hi Alex et al.! How are things going? Let me know if anything is unclear or a misinterpretation from my end. I'd be happy to discuss! :)
