+++ paper/paper.bib
@@ -309,3 +309,11 @@ year = {2021}
   year={2017},
   publisher={eLife Sciences Publications Limited}
 }
+
+@article{Droettboom13,
+  title={airspeed velocity},
+  author={Michael Droettboom and Pauli Virtanen},
+  journal={GitHub},
+  year={2013},
+  url={https://github.com/airspeed-velocity/asv}
+}
