### 2022-06-02 ed59ed0
clarify anndata vs AnnData notation
