+++ paper/paper.bib
@@ -317,3 +317,11 @@ year = {2021}
   year={2013},
   url={https://github.com/airspeed-velocity/asv}
 }
+
+@article{Cannoodt21,
+  title={anndata: 'anndata' for R},
+  author={Robrecht Cannoodt},
+  journal={CRAN},
+  year={2021},
+  url={https://CRAN.R-project.org/package=anndata}
+}
