+++ paper/paper.bib
@@ -325,3 +325,11 @@ year = {2021}
   year={2021},
   url={https://CRAN.R-project.org/package=anndata}
 }
+
+@article{Zappia22,
+  title={Conversion Between scRNA-seq Objects},
+  author={Luke Zappia and Aaron Lun},
+  journal={Bioconductor},
+  year={2022},
+  doi={10.18129/B9.bioc.zellkonverter}
+}
