### 2022-06-02 b4684e3
cite NeurIPS competition
