### 2022-06-02 5348323
Address Robrecht's JOSS review of the anndata paper (#779)
* fixed theislab -> scverse

* mention benchmarks

* add a reference for python adoption

* clarify anndata vs AnnData notation

* fix wickham14 inline citation

* added referenced to anndata CRAN package

* cite zellkonverter

* cite NeurIPS competition
