+++ paper/paper.bib
@@ -309,3 +309,35 @@ year = {2021}
   year={2017},
   publisher={eLife Sciences Publications Limited}
 }
+
+@article{Droettboom13,
+  title={airspeed velocity},
+  author={Michael Droettboom and Pauli Virtanen},
+  journal={GitHub},
+  year={2013},
+  url={https://github.com/airspeed-velocity/asv}
+}
+
+@article{Cannoodt21,
+  title={anndata: 'anndata' for R},
+  author={Robrecht Cannoodt},
+  journal={CRAN},
+  year={2021},
+  url={https://CRAN.R-project.org/package=anndata}
+}
+
+@article{Zappia22,
+  title={Conversion Between scRNA-seq Objects},
+  author={Luke Zappia and Aaron Lun},
+  journal={Bioconductor},
+  year={2022},
+  doi={10.18129/B9.bioc.zellkonverter}
+}
+
+@inproceedings{Luecken21,
+  title={A sandbox for prediction and integration of {DNA}, {RNA}, and proteins in single cells},
+  author={Malte D Luecken and Daniel Bernard Burkhardt and Robrecht Cannoodt and Christopher Lance and Aditi Agrawal and Hananeh Aliee and Ann T Chen and Louise Deconinck and Angela M Detweiler and Alejandro A Granados and Shelly Huynh and Laura Isacco and Yang Joon Kim and Dominik Klein and BONY DE KUMAR and Sunil Kuppasani and Heiko Lickert and Aaron McGeever and Honey Mekonen and Joaquin Caceres Melgarejo and Maurizio Morri and Michaela M{\"u}ller and Norma Neff and Sheryl Paul and Bastian Rieck and Kaylie Schneider and Scott Steelman and Michael Sterr and Daniel J. Treacy and Alexander Tong and Alexandra-Chloe Villani and Guilin Wang and Jia Yan and Ce Zhang and Angela Oliveira Pisco and Smita Krishnaswamy and Fabian J Theis and Jonathan M. Bloom},
+  booktitle={Thirty-fifth Conference on Neural Information Processing Systems Datasets and Benchmarks Track (Round 2)},
+  year={2021},
+  url={https://openreview.net/forum?id=gN35BGa1Rt}
+}
