### 2022-06-07 scverse/anndata#769 @rcannood
Thanks for your hard work @falexwolf and colleagues! I ticked off the check boxes for which I have no further questions or comments.

* General issues → Repository
  > We fixed the repository link! https://github.com/scverse/anndata/commit/d409c48ec8ca26707f149fcd23cffbdf970d9f5b

  The first issue at https://github.com/openjournals/joss-reviews/issues/4371 still lists github.com/theislab/anndata instead of scverse/anndata. I think it's important that this is fixed, because the url in the JOSS article itself also still points to the theislab organisation.

* Functionality → Performance
  > It makes a comparison with the loom on-disk format here: https://anndata.readthedocs.io/en/latest/benchmark-read-write.html

  The read/write benchmark is a good start! This only covers a very specific aspect of AnnData and only compares to loom. I'm not sure I'm wholly satisfied that this is sufficient to substantiate your claims regarding performance. The manuscript states: "Due to the increasing scale of data, we emphasized efficient operations with low memory and runtime overhead."
    - "Increasing scale of data" → the dataset used in this benchmark only contains 2300 cells
    - "efficient operations" → the same paragraph lists these operations:
      * out of core conversions between dense and sparse data
      * lazy subsetting
      * per-element operations for low total memory usage
      * in-place subsetting
      * combining `AnnData` objects with various merge strategies
      * lazy concatenation
      * batching
      * backed out-of-memory mode
    - "with low memory [...] overhead" → the benchmark currently doesn't cover memory overhead.
    - "low [...] runtime overhead" → the current benchmark compares against loom to demonstrate runtime efficiency. I wonder, how does it stack up against SCE when performing similar operations?

* Functionality → Performance
  > It links to https://github.com/ivirshup/anndata-benchmarks, an entire devoted to benchmarking anndata over its entire lifetime using [airspeed velocity](https://asv.readthedocs.io/en/stable/).

  It's very cool that you're able to track the performance of anndata in this way! However, I'm having a hard time browsing the results from this workflow. Are the results of the ASV benchmarks the one that are published in readthedocs?

* Functionality → Performance
  > We will add a more comprehensive comparative benchmarks in the docs, link it from the landing page and maintain it.

  Let me know when I can take a look at the extended benchmarks. I realise that the above-mentioned comments regarding the benchmarks are quite extensive. Maybe we can ask @luizirber to chime in on to what extent each and every performance-related claim in the paper needs to be backed by benchmarks?

* Documentation → Installation
  > We agree with you that a more comprehensive installation page in the docs could be beneficial! We will add this page similar to what we set up for Scanpy.

  Sounds good! Let me know when you'd like to me to take a look at that.

* Quality of writing → figure text in figure 1
  > Agreed on this! We will remake the figure. Would you think the figure is acceptable if we increase the fontsize of “obs_names” and “var_names”? Or do you think we have to increase the fontsize everywhere?

  Increasing the size of just the `obs_names`, `var_names` should be sufficient :)
