+++ anndata/_core/anndata.py
@@ -57,7 +57,7 @@ from ..compat import (
     _overloaded_uns,
     OverloadedDict,
 )
-
+from .formatting_html import _create_anndata_repr
 
 class StorageType(Enum):
     Array = np.ndarray
@@ -589,6 +589,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             return "View of " + self._gen_repr(self.n_obs, self.n_vars)
         else:
             return self._gen_repr(self.n_obs, self.n_vars)
+
+    def _repr_html_(self) -> str:
+        """
+        HTML formatter for the AnnData object
+        for rich display in notebooks.
+        """
+        return _create_anndata_repr(self)
 
     def __eq__(self, other):
         """Equality testing"""
