+++ anndata/_core/formatting_html.py
@@ -1,0 +1,422 @@
+"""\
+Utility functions for AnnData._repr_html_()
+"""
+import warnings
+import collections.abc as cabc
+from collections import OrderedDict
+from copy import copy, deepcopy
+from enum import Enum
+from functools import partial, singledispatch
+from pathlib import Path
+from os import PathLike
+from textwrap import dedent
+from typing import Any, Union, Optional  # Meta
+from typing import Iterable, Sequence, Mapping, MutableMapping  # Generic ABCs
+from typing import Tuple, List  # Generic
+
+import numpy as np
+from numpy import ma
+import pandas as pd
+from pandas.api.types import infer_dtype, is_string_dtype, is_categorical_dtype
+from scipy import sparse
+from scipy.sparse import issparse, csr_matrix
+
+from anndata._warnings import ImplicitModificationWarning
+from .raw import Raw
+from .index import _normalize_indices, _subset, Index, Index1D, get_vector
+from .file_backing import AnnDataFileManager, to_memory
+from .access import ElementRef
+from .aligned_mapping import (
+    AxisArrays,
+    AxisArraysView,
+    PairwiseArrays,
+    PairwiseArraysView,
+    Layers,
+    LayersView,
+)
+from .views import (
+    ArrayView,
+    DictView,
+    DataFrameView,
+    as_view,
+    _resolve_idxs,
+)
+from .sparse_dataset import SparseDataset
+from .. import utils
+from ..utils import convert_to_dict, ensure_df_homogeneous
+from ..logging import anndata_logger as logger
+from ..compat import (
+    ZarrArray,
+    ZappyArray,
+    DaskArray,
+    Literal,
+    _slice_uns_sparse_matrices,
+    _move_adj_mtx,
+    _overloaded_uns,
+    OverloadedDict,
+)
+
+from typing import Tuple, Iterable
+from functools import singledispatch
+from numbers import Integral, Real, Complex
+from warnings import warn
+import numpy as np
+import pandas as pd
+from functools import singledispatch
+from html import escape
+import uuid
+import contextlib
+import uuid
+from collections import OrderedDict
+from functools import lru_cache, partial
+from html import escape
+from importlib.resources import read_binary
+from typing import Collection, Hashable, Literal, TypedDict
+import numpy as np
+
+
+
+STATIC_FILES = (
+    ("anndata.static.html", "icons-svg-inline.html"),
+    ("anndata.static.css", "style.css"),
+)
+
+Options = Literal[
+    "display_max_rows",
+    "display_values_threshold",
+    "display_style",
+    "display_width",
+    "display_expand_attrs",
+    "display_expand_coords",
+    "display_expand_data_vars",
+    "display_expand_data",
+]
+
+class T_Options(TypedDict):
+        display_max_rows: int
+        display_values_threshold: int
+        display_style: Literal["text", "html"]
+        display_width: int
+        display_expand_attrs: Literal["default", True, False]
+        display_expand_coords: Literal["default", True, False]
+        display_expand_data_vars: Literal["default", True, False]
+        display_expand_data: Literal["default", True, False]
+    
+
+OPTIONS: T_Options = {
+    "display_max_rows": 12,
+    "display_values_threshold": 200,
+    "display_style": "html",
+    "display_width": 100,
+    "display_expand_attrs": "default",
+    "display_expand_coords": "default",
+    "display_expand_data_vars": "default",
+    "display_expand_data": "default",
+}
+
+
+
+@lru_cache(None)
+def _load_static_files():
+    """Lazily load the resource files into memory the first time they are needed"""
+    return [
[…]
