+++ anndata/_core/formatting_html.py
@@ -75,7 +75,6 @@ from typing import Collection, Hashable, Literal, TypedDict
 import numpy as np
 
 
-
 STATIC_FILES = (
     ("anndata.static.html", "icons-svg-inline.html"),
     ("anndata.static.css", "style.css"),
@@ -92,16 +91,17 @@ Options = Literal[
     "display_expand_data",
 ]
 
+
 class T_Options(TypedDict):
-        display_max_rows: int
-        display_values_threshold: int
-        display_style: Literal["text", "html"]
-        display_width: int
-        display_expand_attrs: Literal["default", True, False]
-        display_expand_coords: Literal["default", True, False]
-        display_expand_data_vars: Literal["default", True, False]
-        display_expand_data: Literal["default", True, False]
-    
+    display_max_rows: int
+    display_values_threshold: int
+    display_style: Literal["text", "html"]
+    display_width: int
+    display_expand_attrs: Literal["default", True, False]
+    display_expand_coords: Literal["default", True, False]
+    display_expand_data_vars: Literal["default", True, False]
+    display_expand_data: Literal["default", True, False]
+
 
 OPTIONS: T_Options = {
     "display_max_rows": 12,
@@ -113,7 +113,6 @@ OPTIONS: T_Options = {
     "display_expand_data_vars": "default",
     "display_expand_data": "default",
 }
-
 
 
 @lru_cache(None)
@@ -129,9 +128,11 @@ def _load_static_files():
 def _html_format(x):
     return escape(str(x))
 
+
 @_html_format.register(pd.DataFrame)
 def _html_format_df(x: pd.DataFrame):
     return x._repr_html_()
+
 
 @_html_format.register(np.ndarray)
 def _html_format_np(x: np.ndarray):
@@ -152,6 +153,7 @@ def _html_format_np(x: np.ndarray):
     with set_numpy_options(**options):
         return repr(x)
 
+
 @_html_format.register(sparse.spmatrix)
 def _html_format_sp(x: sparse.spmatrix):
     return f"<pre>{escape(x.__repr__())}</pre>"
@@ -161,9 +163,11 @@ def _html_format_sp(x: sparse.spmatrix):
 def _dtype_repr(x):
     return escape(str(x.dtype))
 
+
 @_dtype_repr.register(pd.DataFrame)
 def _dtype_repr_pd(x: pd.DataFrame):
     return f"DataFrame {escape(str(x.shape))}"
+
 
 # TODO: document and mention
 # https://github.com/pydata/xarray/blob/8e9a9fb390f8f0d27a017a7affd8d308d2317959/xarray/core/formatting.py#L32
@@ -176,8 +180,7 @@ def maybe_truncate(obj, max_width=500):
 
 @singledispatch
 def _inline_format(x, max_width):
-    return maybe_truncate(escape(str(x)),max_width=max_width)
-
+    return maybe_truncate(escape(str(x)), max_width=max_width)
 
 
 @contextlib.contextmanager
@@ -211,6 +214,7 @@ def _obj_repr(obj, header_components, sections):
         "</div>"
     )
 
+
 def _summarize_attrs(attrs):
     attrs_dl = "".join(
         f"<dt><span>{escape(str(k))} :</span></dt>" f"<dd>{escape(str(v))}</dd>"
@@ -231,10 +235,12 @@ def _icon(icon_name):
     )
 
 
-def _summarize_item_html(name:str, x:Union[pd.DataFrame,np.ndarray,sparse.spmatrix],attrs=None):
+def _summarize_item_html(
+    name: str, x: Union[pd.DataFrame, np.ndarray, sparse.spmatrix], attrs=None
+):
     """
     Summarizes x
-    """    
+    """
[…]
