### 2022-06-15 ac2abc9
now collapsing can be configured
