### 2022-06-19 #4371 @falexwolf
@arfon We moved the repository from theislab/anndata to https://github.com/scverse/anndata. Could you update the metadata here?
