### 2022-06-19 scverse/anndata#769 @falexwolf
Thanks a lot for this second round, @rcannood! (Why am I so slow in responding? Sorry!)

I pinged the editor re the repo location and am talking to @ivirshup about the rest!
