### 2022-06-20 #4371 @danielskatz
editorialbot commands need to be the first entry in a new comment.
