### 2022-06-20 scverse/anndata#769 @ivirshup
Hey @rcannood, sorry also missed the update here!

> "Increasing scale of data" → the dataset used in this benchmark only contains 2300 cells

Looking into making something more up to date here!

> It's very cool that you're able to track the performance of anndata in this way! However, I'm having a hard time browsing the results from this workflow. Are the results of the ASV benchmarks the one that are published in readthedocs?

They are not.

At the moment, we just treat this as tooling for development, and don't publish results. We need to get dedicated hardware set up to have meaningful historical results (since performance can be quite variable between machines otherwise).

Additionally, the benchmarks here can be quite specific, and I'm not sure they would be informative by themselves. Here is an example of output: https://www.astropy.org/astropy-benchmarks/
