### 2022-06-21 348af11
some tuning and tests
