### 2022-06-21 scverse/anndata#769 @nlhepler
Hello all!

Thank you for the very well-written paper and similarly well-built and well-maintained anndata package. I concur with @rcannood that it is worthy of publication, after a few things have been addressed:

- It's already been noted, but the package has since moved to scverse/anndata. I presume this will get fixed in time for publication, but it's also not necessarily blocking as an auto-redirect is in place.
- I'm not sure if it's because the figures possibly appeared first in poster form, but the figure text is far too small for the 8.5x11" format. In particular Figure 1 and Figure 3 need to have the text resized to be at least 10pt (including/especially the TeX bits). Figure 2 could also use some massaging, but its text is not nearly as egregiously tiny.
- The data structure section is missing a description/enumeration of the layers functionality of the format. This is described in the Figure 1 caption but absent from the main text.
- The same section is also missing a description of `uns`, which is also present in Figure 1's caption. It is possibly more important to describe than layers as you refer to `uns` on line 132 of the main text.
- I have a question: why are single- and multi-dimensional annotations separated into `var` and `obs` vs `varm` and `obsm` respectively? A single-dimensional annotation has an implicit dimension of 1 and can leverage the same datastructure as annotations with dimension >1?
- The data analysis workflow section, line 74: "kept track of~f~ by adding"
- The on-disk format section, line 108: say "HDF5 AnnData format" instead of .h5ad, or parenthetically define .h5ad and .zarr on e.g. line 101.
- Another question: line 108: you say the schema is versioned and stored in an internal registry. Is this registry internal to every file, as in a self-describing format, or internal to the anndata project? Either answer is acceptive, it may help to clarify in the paper.

A couple remarks I think warrant some interaction/debate:
- I think it is maybe inaccurate to imply AnnData is a datastructure library (The AnnData object section). It seems to leverage existing datastructure libraries (for integration purposes, like numpy, scipy.sparse, and pandas), and provides a layer of APIs on top of these datastructures that render programmatic management of experimental data of this form easy to manipulate, integrate with downstream analyses, and serialize to and from disk.
- Related to the above, Between the Introduction and the first sentence of The AnnData object section there's an indirect statement that AnnData is fulfilling an important niche: providing ExpressionSet-like functionality for Python language users, and integrating with the relevant packages in the Python ecosystem to provide access to e.g. modern machine learning and other (scientific) software unique to the Python ecosystem. It may be better to directly state this.

I also have a few other comments that I would like to share that (I do not think block publication but) might be helpful if you would like to further polish this submission or if you intend to submit a follow-up paper in the future:

- AnnData is relatively mature, having developed originally as part of the scanpy project. As such, I think the Outlook section (or another) could contain a discussion of lessons learned and/or resolved, in-scope for future development efforts, or out-of-scope but something to think about in any rewrite/successor project. Comparisons with prior versions could also be a source of benchmarks.
- Most papers describing new software would have benchmarks showcasing absolute performance claims. The paper as-is describes its performance qualitatively, rather than in quantitative terms, which is less useful/informative, but does indicate that the authors are aware how important time/space efficiency are to practical use of the project. Unfortunately, AnnData does not have any serious competitors in the Python ecosystem, and the R-based competitors are ... well it can often be a totally different set of trade-offs. So "advancing state of the art" here would admittedly be difficult to showcase. Instead it might be more appropriate to quantify the amount of boilerplate saved by using AnnData over rolling one's own data loaders/accessors, and to extol more assertively the benefits of standardized on-disk formats.
- Related to the above, I would appreciate any future papers on AnnData to engage in self-comparative benchmarks to showcase the advantages of any new features/APIs/developments/etc.
- As someone who has worked on similar software in a corporate setting for highly efficient storage and high-performance analysis of large datasets (100K+ cells) on commodity PC hardware, I know how important it is to provide low-level APIs that can be leveraged to write highly efficient applications. This is especially true in the context of lazy subsetting where you really do not want to fully instantiate a sub-matrix and possibly double your memory usage and do lots of unnecessary copying. I would really appreciate a detailed treatment of any such APIs in this or any future paper. In fact, I would argue that such work is what has made numpy such an important and successful project, as low-level high-performance APIs made it feasible to build many downstream packages (e.g. scipy, sklearn, etc) and integrations (e.g. Cython).

Thank you all again for the opportunity to review this paper and render remarks!

Best,

Lance Hepler
