### 2022-06-21 8269f73
Update rename_categories internals
