+++ anndata/tests/test_base.py
@@ -452,10 +452,14 @@ def test_rename_categories():
     adata.uns["tool"]["params"] = dict(groupby="cat_anno")
 
     new_categories = ["c", "d"]
-    adata.rename_categories("cat_anno", new_categories)
+    with pytest.warns(None) as records:
+        adata.rename_categories("cat_anno", new_categories)
 
     assert list(adata.obs["cat_anno"].cat.categories) == new_categories
     assert list(adata.uns["tool"]["cat_array"].dtype.names) == new_categories
+
+    if len(records) > 0:
+        raise AssertionError(records)
 
 
 def test_pickle():
