### 2022-06-21 9b3ba2e
Update how we check no warnings are thrown
