+++ anndata/tests/test_base.py
@@ -1,4 +1,5 @@
 from itertools import product
+import warnings
 
 import numpy as np
 from numpy import ma
@@ -452,14 +453,12 @@ def test_rename_categories():
     adata.uns["tool"]["params"] = dict(groupby="cat_anno")
 
     new_categories = ["c", "d"]
-    with pytest.warns(None) as records:
+    with warnings.catch_warnings():
+        warnings.simplefilter("error")
         adata.rename_categories("cat_anno", new_categories)
 
     assert list(adata.obs["cat_anno"].cat.categories) == new_categories
     assert list(adata.uns["tool"]["cat_array"].dtype.names) == new_categories
-
-    if len(records) > 0:
-        raise AssertionError(records)
 
 
 def test_pickle():
