### 2022-06-21 869a8c9
Backport PR #790: Update rename_categories internals
