### 2022-06-24 d6da02f
Attribute update and some tests
