+++ anndata/_core/formatting_html.py
@@ -6,7 +6,6 @@ import sys
 import uuid
 from functools import lru_cache, singledispatch
 from html import escape
-import contextlib
 from typing import Union, Mapping
 from importlib.resources import read_binary
 from scipy import sparse
@@ -82,6 +81,21 @@ def _load_static_files():
     ]
 
 
+# TODO: document and mention
+# https://github.com/pydata/xarray/blob/8e9a9fb390f8f0d27a017a7affd8d308d2317959/xarray/core/formatting.py#L32
+def _maybe_truncate(obj, max_width=500):
+    s = str(obj)
+    if len(s) > max_width:
+        s = s[: (max_width - 3)] + "..."
+    return s
+
+
+"""
+HTML FORMAT FUNCTIONS
+These functions give an representation to be shown in the html summary.
+"""
+
+
 @singledispatch
 def _html_format(x):
     return f"<pre>{escape(str(x))}</pre>"
@@ -117,20 +131,11 @@ def _html_format_sp(x: sparse.spmatrix):
     return f"<pre>{escape(x.__repr__())}</pre>"
 
 
-# TODO: find a test case when the array variable enters the last branch
-# then implement a similar mechanism for repr.
-# def short_data_repr(array):
-#     """Format "data" for DataArray and Variable."""
-#     internal_data = getattr(array, "variable", array)._data
-#     if isinstance(array, np.ndarray):
-#         return short_numpy_repr(array)
-#     elif is_duck_array(internal_data):
-#         return limit_lines(repr(array.data), limit=40)
-#     elif array._in_memory or array.size < 1e5:
-#         return short_numpy_repr(array)
-#     else:
-#         # internal xarray array type
-#         return f"[{array.size} values with dtype={array.dtype}]"
+"""
+Dimension repr FUNCTIONS
+Gives the shape of different data-structures based on their type.
+"""
+
 
 @singledispatch
 def _dim_repr(x):
@@ -144,6 +149,12 @@ def _dim_repr(x):
 @_dim_repr.register(np.ndarray)
 def _dim_repr_shape(x):
     return f"({', '.join(escape(str(dim)) for dim in x.shape)})"
+
+
+"""
+Type repr FUNCTIONS
+Gives a short representation of the type of the data structure.
+"""
 
 
 @singledispatch
@@ -164,18 +175,44 @@ def _type_repr_pd(x: pd.DataFrame):
     return f"DataFrame {escape(str(x.shape))}"
 
 
-# TODO: document and mention
-# https://github.com/pydata/xarray/blob/8e9a9fb390f8f0d27a017a7affd8d308d2317959/xarray/core/formatting.py#L32
-def maybe_truncate(obj, max_width=500):
-    s = str(obj)
-    if len(s) > max_width:
-        s = s[: (max_width - 3)] + "..."
-    return s
+"""
+Default Attribute FUNCTIONS
+Adds attributes depending on the type of the data structure.
+By default does nothing but if overloaded, it can add attributes to the specific data
+type automatically.
+"""
+
+
+@singledispatch
+def _add_attrs(x, attrs:dict):
+    return attrs
+
+
+@_add_attrs.register(pd.DataFrame)
+def _add_attrs_pd_df(x : pd.DataFrame, attrs):
+    for c,t in zip(x.columns,x.dtypes):
+        txt = None
+        if isinstance(t,pd.CategoricalDtype):
+            txt = "Ordered " if t.ordered else "Unordered "
+            txt += "Cat.: " + ", ".join(t.categories)
+            txt = escape(txt)
+        else:
+            txt = escape(str(t))
+        key = escape("col_" + str(c) + "_dtype")
+        if attrs.get(key) is None:
+            attrs[key] = txt
+    return attrs
+
+
+"""
[…]
