### 2022-07-07 a61ba83
added pandas config and vscode theme looks better
