### 2022-07-13 940a7e6
bugfix, todo: truncate categories
