### 2022-07-19 fee56ee
Update dim_len
