### 2022-07-19 4b89a9b
Validate that awkward arrays align to axes
