### 2022-07-19 9d56157
Fix reindexing during merge
