### 2022-07-19 c4c1b3f
Test different types of awkward arrays in different slots
