### 2022-07-19 339bce8
Better function to generate awkward arrays
