### 2022-07-19 012de5e
Better dim_len for awkward arrays
