### 2022-07-19 7884598
Working out how to best check the dim_len
