### 2022-07-20 e16ae35
Only accept awkward arrays that are "regular" in the aligned dimension
The conversion is left to the user. Explicit is better than implicit.
