### 2022-08-05 #4371 @nlhepler
:wave: Hi @arfon, the conversation is mostly happening in scverse/anndata#769 --
