### 2022-08-08 e2d4bb3
Suport buffers. TODO: tests
