### 2022-08-08 25893e3
Added error tests
