### 2022-08-08 839ec42
Define protocols ourselves
