### 2022-08-10 7180c17
H5Py supports file objects
