+++ anndata/tests/test_hdf5_backing.py
@@ -313,21 +313,6 @@ def test_backed_modification_sparse(adata, backing_h5ad, sparse_format):
     assert np.all(adata.X[2, :] == np.array([7, 0, 9]))
 
 
-def test_backed_no_buffer(adata, backing_h5ad):
-    buf = io.BytesIO()
-    with pytest.raises(
-        TypeError, match=r"^Can’t use backed mode with a file-like object\.$"
-    ):
-        ad.read_h5ad(buf, backed=True)
-
-    adata.write(backing_h5ad)
-    adata = ad.read_h5ad(backing_h5ad, backed="r+")
-    with pytest.raises(
-        TypeError, match=r"^Can’t write backed AnnData object to file-like object."
-    ):
-        adata.write(buf)
-
-
 # TODO: Work around h5py not supporting this
 # def test_backed_view_modification(adata, backing_h5ad):
 #     adata.write(backing_h5ad)
