+++ anndata/_core/merge.py
@@ -129,7 +129,7 @@ def equal_sparse(a, b) -> bool:
 
 @equal.register(AwkArray)
 def equal_awkward(a, b) -> bool:
-    import awkward as ak
+    import awkward._v2 as ak
 
     if dim_len(a, 0) == dim_len(b, 0):
         return ak.all(a == b)
@@ -465,7 +465,7 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
         df.index = index
         return df
     elif any(isinstance(a, AwkArray) for a in arrays):
-        import awkward as ak
+        import awkward._v2 as ak
 
         return ak.concatenate([f(a, axis=1 - axis) for f, a in zip(reindexers, arrays)])
     elif any(isinstance(a, sparse.spmatrix) for a in arrays):
