### 2022-08-12 e687e19
WIP rewrite awkward array generation
