### 2022-08-12 41b1423
Improve awkward array generation and dim_len check
