### 2022-08-12 fa8a386
Switch to new awkward array generation in all tests
