### 2022-08-12 733937a
Fix test_transpose
