### 2022-08-13 e5706c3
Add test for setting anndata slots to awkward arrays
