### 2022-08-13 fceab1b
enable tests for 3d ragged array in layers
