### 2022-08-13 06608e9
Fix that X could not be set when creating AnnData object from scratch.
Apparently the checks are quite different than when adding a Layer.
