### 2022-08-19 3c46363
Remove code to make awkward array regular after merge.
This is now done by the awkward array library.
