### 2022-08-29 285e3b3
Do not explicitly copy awkward arrays
