+++ anndata/_core/views.py
@@ -122,10 +122,8 @@ class DataFrameView(_ViewMixin, pd.DataFrame):
 
 class AwkwardArrayView(_ViewMixin, AwkArray):
     def copy(self, order: str = "C") -> np.ndarray:
-        import awkward._v2 as ak
-
-        # we want a copy of an akward array
-        return ak.copy(self)
+        # awkward arrays are immutable, we don't need to make an explicit copy.
+        return self
 
 
 @singledispatch
