### 2022-08-29 32c44cf
Add docs stub and update type hints
