### 2022-08-29 5e1c1da
Fix: dtype not available during merge if both X are awkward
