+++ pyproject.toml
@@ -90,7 +90,7 @@ test = [
     "boltons",
     "scanpy",
     "dask[array]",
-    "awkward>=1,<2",
+    "awkward>=1.9.0rc12,<2",
 ]
 
 [tool.flit.sdist]
