### 2022-08-30 2e66409
Exclude awkward layer in loom tests
