+++ anndata/compat/__init__.py
@@ -11,7 +11,6 @@ import numpy as np
 import pandas as pd
 
 from ._overloaded_dict import _overloaded_uns, OverloadedDict
-from .._core.index import _subset
 
 
 class Empty:
@@ -39,6 +38,16 @@ except ImportError:
         @staticmethod
         def __repr__():
             return "mock zarr.core.Group"
+
+
+try:
+    from awkward._v2 import Array as AwkArray
+except ImportError:
+
+    class AwkArray:
+        @staticmethod
+        def __repr__():
+            return "mock awkward.highlevel.Array"
 
 
 try:
@@ -248,6 +257,8 @@ def _find_sparse_matrices(d: Mapping, n: int, keys: tuple, paths: list):
 
 
 def _slice_uns_sparse_matrices(uns: MutableMapping, oidx: "Index1d", orig_n_obs: int):
+    from anndata._core.index import _subset
+
     """slice sparse spatrices of n_obs × n_obs in self.uns"""
     if isinstance(oidx, slice) and len(range(*oidx.indices(orig_n_obs))) == orig_n_obs:
         return uns  # slice of entire dimension is a no-op
