+++ pyproject.toml
@@ -76,6 +76,7 @@ doc = [
     "scanpydoc>=0.7.7",
     "typing_extensions; python_version < '3.8'",
     "zarr",
+    "awkward>=1.9.0rc12,<2",
 ]
 test = [
     "loompy>=3.0.5",
@@ -89,6 +90,7 @@ test = [
     "boltons",
     "scanpy",
     "dask[array]",
+    "awkward>=1.9.0rc12,<2",
 ]
 
 [tool.flit.sdist]

[…]
