+++ anndata/tests/helpers.py
@@ -11,6 +11,7 @@ from pandas.api.types import is_numeric_dtype
 import pytest
 from scipy import sparse
 import awkward._v2 as ak
+import random
 
 from anndata import AnnData, Raw
 from anndata._core.views import ArrayView
