+++ anndata/_core/anndata.py
@@ -269,8 +269,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         obs: Optional[Union[pd.DataFrame, Mapping[str, Iterable[Any]]]] = None,
         var: Optional[Union[pd.DataFrame, Mapping[str, Iterable[Any]]]] = None,
         uns: Optional[Mapping[str, Any]] = None,
-        obsm: Optional[Union[np.ndarray, Mapping[str, Sequence[Any], AwkArray]]] = None,
-        varm: Optional[Union[np.ndarray, Mapping[str, Sequence[Any], AwkArray]]] = None,
+        obsm: Optional[Union[np.ndarray, Mapping[str, Sequence[Any]]]] = None,
+        varm: Optional[Union[np.ndarray, Mapping[str, Sequence[Any]]]] = None,
         layers: Optional[Mapping[str, Union[np.ndarray, sparse.spmatrix]]] = None,
         raw: Optional[Mapping[str, Any]] = None,
         dtype: Optional[Union[np.dtype, type, str]] = None,
