+++ docs/release-notes/0.8.1.rst
@@ -4,5 +4,6 @@
 .. rubric:: Bug fixes
 
 * Fix warning from `rename_categories` :pr:`790` :smaller:`I Virshup`
+* Remove backwards compat checks for categories in `uns` when we can tell the file is new enough :pr:`790` :smaller:`I Virshup`
 
 .. rubric:: Documentation
