### 2022-08-30 7fc7d9b
Add a test to ensure equivalent ordered categoricals aren't cast to object
