### 2022-08-30 5cb1911
Implement concat with category merging (#763)
* Implement concat with category merging

* fix order

* Simplify computation

* Check behaviour for outer joins and ordered categoricals

* some internal docs

* Add a test to ensure equivalent ordered categoricals aren't cast to object

* release note
