### 2022-08-31 4922603
Update error message in algined mapping
