### 2022-08-31 c5c5335
Use compat module to support both awkward v1.9rc and 2.x
