+++ anndata/compat/__init__.py
@@ -41,7 +41,21 @@ except ImportError:
 
 
 try:
-    from awkward._v2 import Array as AwkArray
+    try:
+        from importlib.metadata import version
+    except ImportError:
+        from importlib_metadata import version
+
+    # Ensure compatibility with both 1.9rc and 2.x releases of awkward
+    # TODO Once 2.x is released, we can require it as a minimal dependency and remove this code.
+    major, _ = version("awkward").split(".", maxsplit=1)
+    if int(major) < 2:
+        import awkward._v2 as awkward
+    else:
+        import awkward
+
+    AwkArray = awkward.Array
+
 except ImportError:
 
     class AwkArray:
