### 2022-08-31 a8e1648
Add tests for copies and view
