### 2022-09-02 d836999
Fix how actual shape is computed in aligned mapping
