### 2022-09-02 ed95d8f
Attempt to support views with ak.behavior
