+++ anndata/_io/h5ad.py
@@ -184,6 +184,12 @@ def read_h5ad(
         instead of fully loading it into memory (`memory` mode).
         If you want to modify backed attributes of the AnnData object,
         you need to choose `'r+'`.
+
+        Currently, `backed` only support updates to `X`. That means any
+        changes to other slots like `obs` will not be written to disk in
+        `backed` mode. If you would like save changes made to these slots
+        of a `backed` :class:`~anndata.AnnData`, write them to a new file
+        (see :meth:`~anndata.AnnData.write`).
     as_sparse
         If an array was saved as dense, passing its name here will read it as
         a sparse_matrix, by chunk of size `chunk_size`.
