### 2022-09-02 f7edc67
Use shallow copy
