### 2022-09-02 c8dce43
copy paste util func
