### 2022-09-03 5a1d056
Add dim_len_awkward function including tests
