### 2022-09-03 83effad
Test that assigning an awkward v1 arrays fails
