### 2022-09-03 21a4b5f
Add stub for element-wise IO tests
