+++ anndata/utils.py
@@ -72,6 +72,54 @@ def dim_len(x, axis):
 try:
     from .compat import awkward as ak
 
+    def _size_at_depth(layout, depth, lateral_context, **kwargs):
+        """Callback function for dim_len_awkward, resolving the dim_len for a given level"""
+        if layout.is_NumpyType:
+            # if it's an embedded rectilinear array, we have to deal with its shape
+            # which might not be 1-dimensional
+            if layout.is_UnknownType:
+                shape = (0,)
+            else:
+                shape = layout.shape
+            numpy_axis = lateral_context["axis"] - depth + 1
+            if not (1 <= numpy_axis < len(shape)):
+                raise TypeError(f"axis={lateral_context['axis']} is too deep")
+            lateral_context["out"] = shape[numpy_axis]
+            return layout.nplike.empty(1)
+
+        elif layout.is_ListType and depth == lateral_context["axis"]:
+            if layout.is_RegularType:
+                # if it's a regular list, you want the size
+                lateral_context["out"] = layout.size
+            else:
+                # if it's an irregular list, you want a null token
+                lateral_context["out"] = -1
+            return layout.nplike.empty(1)
+
+        elif layout.is_RecordType:
+            # if it's a record, you want to stop descent with an error
+            raise TypeError(
+                f"axis={lateral_context['axis']} is too deep, reaches record"
+            )
+
+        elif layout.is_UnionType:
+            # if it's a union, you could get the result of each union branch
+            # separately and see if they're all the same; if not, it's an error
+            result = None
+            for content in layout.contents:
+                context = {"axis": lateral_context["axis"]}
+                ak.transform(
+                    _size_at_depth, content, lateral_context=context, return_array=False
+                )
+                if result is None:
+                    result = context["out"]
+                elif result != context["out"]:
+                    # Union branches have different lengths -> return null token
+                    lateral_context["out"] = -1
+                    return layout.nplike.empty(1)
+            lateral_context["out"] = result
+            return layout.nplike.empty(1)
+
     @dim_len.register(ak.Array)
     def dim_len_awkward(array, axis):
         """Get the length of an awkward array in a given dimension
@@ -86,62 +134,15 @@ try:
             return len(array)
         else:
 
-            def size_at_depth(layout, depth, lateral_context, **kwargs):
-                if layout.is_NumpyType:
-                    # if it's an embedded rectilinear array, we have to deal with its shape
-                    # which might not be 1-dimensional
-                    if layout.is_UnknownType:
-                        shape = (0,)
-                    else:
-                        shape = layout.shape
-                    numpy_axis = lateral_context["axis"] - depth + 1
-                    if not (1 <= numpy_axis < len(shape)):
-                        raise TypeError(f"axis={lateral_context['axis']} is too deep")
-                    lateral_context["out"] = shape[numpy_axis]
-                    return layout.nplike.empty(1)
-
-                elif layout.is_ListType and depth == lateral_context["axis"]:
-                    if layout.is_RegularType:
-                        # if it's a regular list, you want the size
-                        lateral_context["out"] = layout.size
-                    else:
-                        # if it's an irregular list, you want a null token
-                        lateral_context["out"] = -1
-                    return layout.nplike.empty(1)
-
-                elif layout.is_RecordType:
-                    # if it's a record, you want to stop descent with an error
-                    raise TypeError(
-                        f"axis={lateral_context['axis']} is too deep, reaches record"
-                    )
-
-                elif layout.is_UnionType:
-                    # if it's a union, you could get the result of each union branch
-                    # separately and see if they're all the same; if not, it's an error
-                    result = None
-                    for content in layout.contents:
-                        context = {"axis": lateral_context["axis"]}
-                        ak.transform(
-                            size_at_depth,
-                            content,
-                            lateral_context=context,
[…]
