### 2022-09-04 2c59b19
Add more test cases for awkward IO
