### 2022-09-04 4dc0826
Simplify awkward array view code
