### 2022-09-05 3ab5646
Use None to remove name from awkward array
