### 2022-09-05 faf7b1c
add some tests to see where it fails
