### 2022-09-05 e473844
add new ignored filenames
