+++ .github/actions/close-stale.yml
@@ -1,0 +1,20 @@
+name: "Close stale issues"
+on:
+  schedule:
+    - cron: "0 2 * * *"
+
+jobs:
+  stale:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/stale@v5
+        with:
+          days-before-issue-stale: 65535 # We don't want to mark issues as stale in this action
+          days-before-close: 14
+          stale-issue-label: stale
+          any-of-labels: "needs info"
+          stale-issue-message: |
+            This issue has been automatically marked as stale because it has not had recent activity.
+            It will be closed if no further activity occurs. Thank you for your contributions!
+          stale-pr-message: null # if not provided, will not mark PRs as stale
+          debug-only: true # enable dry-run, remove when we know from the logs it's working.
