### 2022-09-05 27d3846
Update anndata/tests/test_dask.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
