+++ .github/actions/label-stale.yml
@@ -13,7 +13,7 @@ jobs:
           days-before-pr-stale: -1 # We don't want to mark PRs as stale
           days-before-close: -1 # We don't want to close issues in this action
           stale-issue-label: stale
-          exempt-issue-labels: pinned,bug,enhancement
+          exempt-issue-labels: pinned,enhancement
           stale-issue-message: |
             This issue has been automatically marked as stale because it has not had recent activity.
             Please add a comment if you want to keep the issue open. Thank you for your contributions!
