### 2022-09-05 eb12fcc
add da.Array to helpers.gen_adata
