### 2022-09-06 d2eaf66
Add test for categorical arrays
