+++ .github/actions/label-stale.yml
@@ -2,6 +2,7 @@ name: "Label stale issues"
 on:
   schedule:
     - cron: "30 1 * * *"
+  workflow_dispatch:
 
 jobs:
   stale:
