+++ .github/actions/close-stale.yml
@@ -2,6 +2,7 @@ name: "Close stale issues"
 on:
   schedule:
     - cron: "0 2 * * *"
+  workflow_dispatch:
 
 jobs:
   stale:
