### 2022-09-06 b3678b6
Update anndata/_core/aligned_mapping.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
