### 2022-09-06 d523c89
Update anndata/tests/helpers.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
