### 2022-09-06 222998d
Update awkward tests to use assert_equal with exact=True
