### 2022-09-07 4acf3c8
add also test_obspvarp
