### 2022-09-07 b5c0723
readwrite tests won't be enough for this case prob
