+++ anndata/tests/test_readwrite.py
@@ -8,6 +8,7 @@ import warnings
 import h5py
 import numpy as np
 import pandas as pd
+import dask.array as da
 from pandas.api.types import is_categorical_dtype
 import pytest
 from scipy.sparse import csr_matrix, csc_matrix
@@ -64,6 +65,14 @@ uns_dict = dict(  # unstructured annotation
 )
 
 
+def _darr_from_list(lst):
+    return da.from_array(lst, chunks="auto")
+
+
+# TODO: Find a way to configure chunk sizes
+# TODO: Discuss why this currently works
+
+
 @pytest.fixture(params=[{}, dict(compression="gzip")])
 def dataset_kwargs(request):
     return request.param
@@ -96,7 +105,7 @@ diskfmt2 = diskfmt
 # ------------------------------------------------------------------------------
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
 def test_readwrite_roundtrip(typ, tmp_path, diskfmt, diskfmt2):
     tmpdir = Path(tmp_path)
     pth1 = tmpdir / f"first.{diskfmt}"
@@ -117,7 +126,7 @@ def test_readwrite_roundtrip(typ, tmp_path, diskfmt, diskfmt2):
     assert_equal(adata2, adata1)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
 def test_readwrite_h5ad(typ, dataset_kwargs, backing_h5ad):
     tmpdir = tempfile.TemporaryDirectory()
     tmpdirpth = Path(tmpdir.name)
@@ -154,7 +163,7 @@ def test_readwrite_h5ad(typ, dataset_kwargs, backing_h5ad):
 
 
 @pytest.mark.skipif(not find_spec("zarr"), reason="Zarr is not installed")
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
 def test_readwrite_zarr(typ, tmp_path):
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
@@ -182,7 +191,7 @@ def test_readwrite_zarr(typ, tmp_path):
     assert_equal(adata, adata_src)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
 def test_readwrite_maintain_X_dtype(typ, backing_h5ad):
     X = typ(X_list)
     adata_src = ad.AnnData(X, dtype="int8")
@@ -215,7 +224,7 @@ def test_maintain_layers(rw):
     assert not np.any((orig.layers["sparse"] != curr.layers["sparse"]).toarray())
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
 def test_readwrite_h5ad_one_dimension(typ, backing_h5ad):
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
@@ -226,7 +235,7 @@ def test_readwrite_h5ad_one_dimension(typ, backing_h5ad):
     assert_equal(adata, adata_one)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
[…]
