+++ anndata/tests/test_dask.py
@@ -10,6 +10,40 @@ from anndata.tests.helpers import assert_equal, asarray
 pytest.importorskip("dask.array")
 
 
+@pytest.fixture(
+    params=[
+        [(2000, 1000), (100, 100)],
+        [(200, 100), (100, 100)],
+        [(200, 100), (100, 100)],
+        [(20, 10), (1, 1)],
+        [(20, 10), (1, 1)],
+    ]
+)
+def sizes(request):
+    return request.param
+
+
+@pytest.fixture(params=["h5ad", "zarr"])
+def diskfmt(request):
+    return request.param
+
+
+@pytest.fixture
+def adata(sizes):
+    import numpy as np
+    import dask.array as da
+
+    (M, N), chunks = sizes
+    X = da.random.random((M, N), chunks=chunks)
+    obs = pd.DataFrame(
+        dict(batch=np.array(["a", "b"])[np.random.randint(0, 2, M)]),
+        index=[f"cell{i:03d}" for i in range(M)],
+    )
+    var = pd.DataFrame(index=[f"gene{i:03d}" for i in range(N)])
+
+    return AnnData(X, obs=obs, var=var)
+
+
 def test_dask_X_view():
     import dask.array as da
 
@@ -23,68 +57,39 @@ def test_dask_X_view():
     view.copy()
 
 
-@pytest.fixture(params=["h5ad", "zarr"])
-def diskfmt(request):
-    return request.param
-
-
-def test_dask_write(tmp_path, diskfmt):
-    import dask.array as da
-
+@pytest.fixture
+def rw(adata, tmp_path, diskfmt):
     pth = tmp_path / f"test_write.{diskfmt}"
     write = lambda x, y: getattr(x, f"write_{diskfmt}")(y)
     read = lambda x: getattr(ad, f"read_{diskfmt}")(x)
-    M, N = 50, 30
-    adata = ad.AnnData(
-        obs=pd.DataFrame(index=[f"cell{i:02}" for i in range(M)]),
-        var=pd.DataFrame(index=[f"gene{i:02}" for i in range(N)]),
-    )
-    adata.X = da.random.randint((M, N))
-    write(adata, pth)
-    result = read(pth)
-    assert_equal(adata, result)
+
+    orig = adata
+    write(orig, pth)
+    curr = read(pth)
+    return curr, orig
 
 
-@pytest.mark.parametrize(
-    "shape,chunks",
-    [
-        [(200, 100), (100, 100)],
-        [(200, 100), (100, 100)],
-        [(20, 10), (1, 1)],
-        [(20, 10), (1, 1)],
-    ],
-)
-def test_assign_X(shape, chunks):
+def test_dask_write(tmp_path, diskfmt, rw):
+    orig, curr = rw
+    assert_equal(orig, curr)
+
+
+# TODO: this function can be made more detailed if there is something that doesn't match
+# For example dtype.
+
+
+def test_assign_X(adata):
     """Check if assignment works"""
     import dask.array as da
     import numpy as np
 
-    darr = da.from_array(np.ones(shape), chunks=chunks)
-    adata = AnnData(darr)
+    adata.X = adata.X * 1
+    # This won't work since the setter converts the data as ndarray
     adata_copy = adata.copy()
-    adata.X = -1 * da.from_array(np.ones(shape), chunks=chunks)
-    assert_equal(asarray(adata.X), -1 * np.ones(shape))
-    assert_equal(asarray(adata_copy.X), np.ones(shape))
+
+    adata.X = -1 * da.ones(adata.X.shape, chunks=adata.X.chunksize)
[…]
