### 2022-09-07 fdadd31
passes zarr tests
