### 2022-09-08 6a6657b
Update categorical syntax, add new categorical test
