### 2022-09-08 bebf412
registered dask_arr view. but when will it break?
