+++ anndata/_core/views.py
@@ -11,7 +11,7 @@ from scipy import sparse
 
 from anndata._warnings import ImplicitModificationWarning
 from .access import ElementRef
-from ..compat import ZappyArray
+from ..compat import ZappyArray, DaskArray
 
 
 class _SetItemMixin:
@@ -126,6 +126,7 @@ def as_view(obj, view_args):
 
 
 @as_view.register(np.ndarray)
+@as_view.register(DaskArray)
 def as_view_array(array, view_args):
     return ArrayView(array, view_args=view_args)
 
