### 2022-09-09 5b346c2
added test for assert_eq
