+++ anndata/tests/test_readwrite.py
@@ -104,7 +104,7 @@ diskfmt2 = diskfmt
 # ------------------------------------------------------------------------------
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_arr])
 def test_readwrite_roundtrip(typ, tmp_path, diskfmt, diskfmt2):
     tmpdir = Path(tmp_path)
     pth1 = tmpdir / f"first.{diskfmt}"
@@ -162,7 +162,7 @@ def test_readwrite_h5ad(typ, dataset_kwargs, backing_h5ad):
 
 
 @pytest.mark.skipif(not find_spec("zarr"), reason="Zarr is not installed")
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_arr])
 def test_readwrite_zarr(typ, tmp_path):
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
@@ -190,7 +190,7 @@ def test_readwrite_zarr(typ, tmp_path):
     assert_equal(adata, adata_src)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_arr])
 def test_readwrite_maintain_X_dtype(typ, backing_h5ad):
     X = typ(X_list)
     adata_src = ad.AnnData(X, dtype="int8")
@@ -223,7 +223,7 @@ def test_maintain_layers(rw):
     assert not np.any((orig.layers["sparse"] != curr.layers["sparse"]).toarray())
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_arr])
 def test_readwrite_h5ad_one_dimension(typ, backing_h5ad):
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
@@ -234,7 +234,7 @@ def test_readwrite_h5ad_one_dimension(typ, backing_h5ad):
     assert_equal(adata, adata_one)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_list])
+@pytest.mark.parametrize("typ", [np.array, csr_matrix, _darr_from_arr])
 def test_readwrite_backed(typ, backing_h5ad):
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
