+++ anndata/tests/test_dask.py
@@ -81,13 +81,12 @@ def test_assign_X(adata):
     import dask.array as da
     import numpy as np
 
-    assert type(adata.X) == da.Array
-    adata.X = adata.X * 1
+    adata.X = da.ones(adata.X.shape)
 
     # This won't work since the setter converts the data as ndarray
     adata_copy = adata.copy()
 
-    adata.X = -1 * da.ones(adata.X.shape, chunks=adata.X.chunksize)
+    adata.X = -1 * da.ones(adata.X.shape)
     assert_equal(asarray(adata.X), -1 * np.ones(adata.X.shape))
     assert_equal(asarray(adata_copy.X), np.ones(adata.X.shape))
 
