### 2022-09-13 1e16b1c
added dask array to assert_equal
