### 2022-09-13 bcc7e1e
Remove 3.7 compat
