+++ anndata/compat/__init__.py
@@ -61,23 +61,6 @@ except ImportError:
             return "mock dask.array.core.Array"
 
 
-try:
-    from typing import Literal
-except ImportError:
-    try:
-        from typing_extensions import Literal
-    except ImportError:
-
-        class LiteralMeta(type):
-            def __getitem__(cls, values):
-                if not isinstance(values, tuple):
-                    values = (values,)
-                return type("Literal_", (Literal,), dict(__args__=values))
-
-        class Literal(metaclass=LiteralMeta):
-            pass
-
-
 @singledispatch
 def _read_attr(attrs: Mapping, name: str, default: Optional[Any] = Empty):
     if default is Empty:
