+++ pyproject.toml
@@ -3,7 +3,6 @@ build-backend = "flit_core.buildapi"
 requires = [
     "flit_core >=3.4,<4",
     "setuptools_scm",
-    "importlib_metadata>=0.7; python_version < '3.8'",
 ]
 
 [project]
