### 2022-09-13 919d34c
Bump python version (#820)
* Bump python version

* Fix pyproject.toml

* release note

* Fix release note

* Remove 3.7 compat
