### 2022-09-15 2d2adc1
Update `.obs/.var` lazily on acess
