+++ anndata/_core/anndata.py
@@ -24,6 +24,7 @@ from scipy import sparse
 from scipy.sparse import issparse, csr_matrix
 
 from anndata._warnings import ImplicitModificationWarning
+from .descriptors import DataFrameDescriptor, UnsDescriptor
 from .raw import Raw
 from .index import _normalize_indices, _subset, Index, Index1D, get_vector
 from .file_backing import AnnDataFileManager, to_memory
@@ -38,8 +39,6 @@ from .aligned_mapping import (
 )
 from .views import (
     ArrayView,
-    DictView,
-    DataFrameView,
     as_view,
     _resolve_idxs,
 )
@@ -51,10 +50,7 @@ from ..compat import (
     ZarrArray,
     ZappyArray,
     DaskArray,
-    _slice_uns_sparse_matrices,
     _move_adj_mtx,
-    _overloaded_uns,
-    OverloadedDict,
 )
 
 
@@ -261,6 +257,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         var={"var_names", "col_names", "index"},
     )
 
+    obs = DataFrameDescriptor(attr="obs")
+    var = DataFrameDescriptor(attr="var")
+    uns = UnsDescriptor()
+
     def __init__(
         self,
         X: Optional[Union[np.ndarray, sparse.spmatrix, pd.DataFrame]] = None,
@@ -332,24 +332,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         # the file is the same as of the reference object
         self.file = adata_ref.file
         # views on attributes of adata_ref
-        obs_sub = adata_ref.obs.iloc[oidx]
-        var_sub = adata_ref.var.iloc[vidx]
         self._obsm = adata_ref.obsm._view(self, (oidx,))
         self._varm = adata_ref.varm._view(self, (vidx,))
         self._layers = adata_ref.layers._view(self, (oidx, vidx))
         self._obsp = adata_ref.obsp._view(self, oidx)
         self._varp = adata_ref.varp._view(self, vidx)
-        # Speical case for old neighbors, backwards compat. Remove in anndata 0.8.
-        uns_new = _slice_uns_sparse_matrices(
-            copy(adata_ref._uns), self._oidx, adata_ref.n_obs
-        )
-        # fix categories
-        self._remove_unused_categories(adata_ref.obs, obs_sub, uns_new)
-        self._remove_unused_categories(adata_ref.var, var_sub, uns_new)
         # set attributes
-        self._obs = DataFrameView(obs_sub, view_args=(self, "obs"))
-        self._var = DataFrameView(var_sub, view_args=(self, "var"))
-        self._uns = uns_new
+        self._uns = adata_ref._uns
         self._n_obs = len(self.obs)
         self._n_vars = len(self.var)
 
@@ -849,19 +838,6 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 v.index = value
 
     @property
-    def obs(self) -> pd.DataFrame:
-        """One-dimensional annotation of observations (`pd.DataFrame`)."""
-        return self._obs
-
-    @obs.setter
-    def obs(self, value: pd.DataFrame):
-        self._set_dim_df(value, "obs")
-
-    @obs.deleter
-    def obs(self):
-        self.obs = pd.DataFrame(index=self.obs_names)
-
-    @property
     def obs_names(self) -> pd.Index:
         """Names of observations (alias for `.obs.index`)."""
         return self.obs.index
@@ -872,19 +848,6 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         self._set_dim_index(names, "obs")
 
     @property
-    def var(self) -> pd.DataFrame:
[…]
