### 2022-09-15 f190fd9
Fix debugging print, test
