### 2022-09-19 caeeba3
write fixed-length strings to HDF5 in some circumstances
variable-length strings incur quite some overhead in terms of file size,
so write arrays of fixed-length strings when the length differences
between the elements of the array are small
