### 2022-09-20 2fee839
Update anndata/_io/specs/methods.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
