+++ anndata/_io/specs/methods.py
@@ -321,13 +321,6 @@ def read_array(elem):
     return elem[()]
 
 
-@_REGISTRY.register_read(H5Array, IOSpec("array", "0.2.0"))
-@_REGISTRY.register_read(ZarrArray, IOSpec("array", "0.2.0"))
-@_REGISTRY.register_read(ZarrArray, IOSpec("string-array", "0.2.0"))
-def read_array(elem):
-    return elem[()]
-
-
 @_REGISTRY.register_read_partial(H5Array, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read_partial(ZarrArray, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read_partial(ZarrArray, IOSpec("string-array", "0.2.0"))
