### 2022-09-20 219516b
added write method separately
