### 2022-09-21 870c8ef
Update anndata/tests/test_dask.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
