### 2022-09-21 da35d4f
add more tests and refactoring
