### 2022-09-21 98c752f
added tests and helpers refactoring
