+++ anndata/tests/test_concatenate.py
@@ -17,8 +17,14 @@ from anndata import AnnData, Raw, concat
 from anndata._core.index import _subset
 from anndata._core import merge
 from anndata.tests import helpers
-from anndata.tests.helpers import assert_equal, darr_from_arr_dense, gen_adata
+from anndata.tests.helpers import (
+    assert_equal,
+    darr_from_arr_dense,
+    gen_adata,
+    gen_adata_dask_args,
+)
 from anndata.utils import asarray
+from anndata.compat import DaskArray
 
 
 @singledispatch
@@ -27,10 +33,15 @@ def filled_like(a, fill_value=None):
 
 
 @filled_like.register(np.ndarray)
-def _filled_array(a, fill_value=None):
+def _filled_array_np(a, fill_value=None):
     if fill_value is None:
         fill_value = np.nan
     return np.broadcast_to(fill_value, a.shape)
+
+
+@filled_like.register(DaskArray)
+def _filled_array(a, fill_value=None):
+    return darr_from_arr_dense(_filled_array_np(a, fill_value))
 
 
 @filled_like.register(sparse.spmatrix)
@@ -60,11 +71,26 @@ def make_idx_tuple(idx, axis):
     return tuple(tup)
 
 
+# Will call func(sparse_matrix) so these types should be sparse compatible
+# See array_type_dense if only dense arrays are expected as input.
+@pytest.fixture(
+    params=[asarray, sparse.csr_matrix, sparse.csc_matrix],
+    ids=["np_array", "scipy_csr", "scipy_csc"],
+)
+def array_type(request):
+    return request.param
+
+
+# Give this array_type_dense instead of array_type if
+# the expected result is compared only content wise and
+# array types aren't expected to be the same
+# test_assert(arr1,arr2) = test_assert(arr1.todense(),arr2.todense())
+# test_assert(arr1,arr2) = test_assert(arr1.todense(),arr2.todense())
 @pytest.fixture(
     params=[asarray, sparse.csr_matrix, sparse.csc_matrix, darr_from_arr_dense],
     ids=["np_array", "scipy_csr", "scipy_csc", "dask_array"],
 )
-def array_type(request):
+def array_type_dense(request):
     return request.param
 
 
@@ -138,8 +164,10 @@ def test_concat_interface_errors():
         (lambda x, **kwargs: x[0].concatenate(x[1:], **kwargs), True),
     ],
 )
-def test_concatenate_roundtrip(join_type, array_type, concat_func, backwards_compat):
-    adata = gen_adata((100, 10), X_type=array_type)
+def test_concatenate_roundtrip(
+    join_type, array_type_dense, concat_func, backwards_compat
+):
+    adata = gen_adata((100, 10), X_type=array_type_dense, **gen_adata_dask_args)
 
     remaining = adata.obs_names
     subsets = []
@@ -234,10 +262,10 @@ def test_concatenate_dense():
     assert np.allclose(var_ma.compressed(), var_ma_ref.compressed())
 
 
-def test_concatenate_layers(array_type, join_type):
+def test_concatenate_layers(array_type_dense, join_type):
     adatas = []
     for _ in range(5):
-        a = array_type(sparse.random(100, 200, format="csr"))
+        a = array_type_dense(sparse.random(100, 200, format="csr"))
         adatas.append(AnnData(X=a, layers={"a": a}))
 
     merged = adatas[0].concatenate(adatas[1:], join=join_type)
@@ -392,10 +420,10 @@ def test_concat_annot_join(obsm_adatas, join_type):
     )
 
 
[…]
