### 2022-09-23 16c47f2
Fixed dask vs sparse array comparision test
