+++ anndata/tests/test_dask.py
@@ -103,12 +103,13 @@ def test_assert_equal_dask_arrays():
     assert_equal(c, d)
 
 
-def test_assert_eq():
+def test_assert_equal_dask_sparse_arrays():
+
     import dask.array as da
     from scipy import sparse
 
-    m = 10
-    x = sparse.random(m, m, format="csr", density=0.1)
+    x = sparse.random(10, 10, format="csr", density=0.1)
     y = da.from_array(asarray(x.toarray()))
+
     assert_equal(x, y)
     assert_equal(y, x)
