+++ anndata/tests/test_dask.py
@@ -67,19 +67,20 @@ def test_dask_write(adata, tmp_path, diskfmt):
     adata.obsm["a"] = da.random.random((M, 10))
     adata.obsm["b"] = da.random.random((M, 10))
     adata.varm["a"] = da.random.random((N, 10))
-    
+
     orig = adata
     write(orig, pth)
     curr = read(pth)
-    
+
     with pytest.raises(Exception):
         assert_equal(curr.obsm["a"], curr.obsm["b"])
 
     assert_equal(curr.varm["a"], orig.varm["a"])
     assert_equal(curr.obsm["a"], orig.obsm["a"])
 
-    #assert_equal(asarray(orig), asarray(curr)) # TODO: find way to compare
-    
+    # assert_equal(asarray(orig), asarray(curr)) # TODO: find way to compare
+
+
 def test_assign_X(adata):
     """Check if assignment works"""
     import dask.array as da
