### 2022-09-26 02365a6
Add testcases for dim_len with awkward arrays of strings
