### 2022-09-26 c20cc31
Fix dim_len for arrays of strings
