### 2022-09-26 2473a65
refactored tests only to_memory needs to be done
