+++ anndata/tests/test_readwrite.py
@@ -15,7 +15,7 @@ import zarr
 
 import anndata as ad
 from anndata.utils import asarray
-from anndata.compat import _read_attr
+from anndata.compat import _read_attr, DaskArray
 
 from anndata.tests.helpers import gen_adata, assert_equal, darr_from_arr
 
@@ -141,15 +141,19 @@ def test_readwrite_h5ad(typ, dataset_kwargs, backing_h5ad):
     assert is_categorical_dtype(adata.raw.var["vanno2"])
     pd.testing.assert_frame_equal(adata.obs, adata_src.obs)
     pd.testing.assert_frame_equal(adata.var, adata_src.var)
-    assert np.all(adata.var.index == adata_src.var.index)
+    assert_equal(adata.var.index, adata_src.var.index)
     assert adata.var.index.dtype == adata_src.var.index.dtype
-    assert type(adata.raw.X) is type(adata_src.raw.X)
-    assert type(adata.raw.varm) is type(adata_src.raw.varm)
-    assert np.allclose(asarray(adata.raw.X), asarray(adata_src.raw.X))
+
+    assert isinstance(adata_src.raw.X, (type(adata.raw.X), DaskArray))
+    assert isinstance(
+        adata_src.uns["uns4"]["c"], (type(adata.uns["uns4"]["c"]), DaskArray)
+    )
+    assert isinstance(adata_src.varm, (type(adata.varm), DaskArray))
+
+    assert_equal(adata.raw.X, adata_src.raw.X)
     pd.testing.assert_frame_equal(adata.raw.var, adata_src.raw.var)
     assert isinstance(adata.uns["uns4"]["a"], (int, np.integer))
     assert isinstance(adata_src.uns["uns4"]["a"], (int, np.integer))
-    assert type(adata.uns["uns4"]["c"]) is type(adata_src.uns["uns4"]["c"])
     assert_equal(adata, adata_src)
 
 
@@ -171,14 +175,23 @@ def test_readwrite_zarr(typ, tmp_path):
     assert is_categorical_dtype(adata.raw.var["vanno2"])
     pd.testing.assert_frame_equal(adata.obs, adata_src.obs)
     pd.testing.assert_frame_equal(adata.var, adata_src.var)
-    assert np.all(adata.var.index == adata_src.var.index)
+    assert_equal(adata.var.index, adata_src.var.index)
     assert adata.var.index.dtype == adata_src.var.index.dtype
-    assert type(adata.raw.X) is type(adata_src.raw.X)
-    assert np.allclose(asarray(adata.raw.X), asarray(adata_src.raw.X))
-    assert np.all(adata.raw.var == adata_src.raw.var)
+
+    # Dev. Note:
+    # either load as same type or load the convert DaskArray to array
+    # since we tested if assigned types and loaded types are DaskArray
+    # this would also work if they work
+    assert isinstance(adata_src.raw.X, (type(adata.raw.X), DaskArray))
+    assert isinstance(
+        adata_src.uns["uns4"]["c"], (type(adata.uns["uns4"]["c"]), DaskArray)
+    )
+    assert isinstance(adata_src.varm, (type(adata.varm), DaskArray))
+
+    assert_equal(adata.raw.X, adata_src.raw.X)
+    assert_equal(adata.raw.var, adata_src.raw.var)
     assert isinstance(adata.uns["uns4"]["a"], (int, np.integer))
     assert isinstance(adata_src.uns["uns4"]["a"], (int, np.integer))
-    assert type(adata.uns["uns4"]["c"]) is type(adata_src.uns["uns4"]["c"])
     assert_equal(adata, adata_src)
 
 
