### 2022-09-26 e88c753
removed the compute call
