### 2022-09-28 6ecbdfb
Added assert_equal
