### 2022-09-29 2813dda
updated the equal function but doesnt give warning rn
