### 2022-09-29 9177cdc
added to_memory test
