### 2022-10-01 #4371 @rcannood
I made my suggestions in https://github.com/scverse/anndata/issues/769, waiting for the authors to get back to us.
