### 2022-10-02 scverse/anndata#769 @falexwolf
@ivirshup are scheduled to meet today for the first time after 2 months. We split up the work. I'll add my responses here for now and very much hope we can finalize with another comment in this thread later today.

I'll address Lance's points in the same order and grouping as he provided them.

Regarding the text:

1. "repo location": Fixed in all places where we could.
2. "figure text": @ivirshup, next comment.
3. "layers": https://github.com/scverse/anndata/pull/825/commits/15160220279e9d297824e24a2ebea8937521c437
4. "uns": https://github.com/scverse/anndata/pull/825/commits/e9a5b02ff5b69f13ef8007d9f6391f24206dc9e2
5. "Why are one- and multi-dimensional annotations separated in two fields?": Two reasons are
   a. Usage patterns: One-dimensional annotations are used differently than multi-dimensional annotations. They always serve as labels with some semantic content (even if numeric), while multi-dimensional annotations store representations of the data that need additional algorithms for interpretation (assigning semantic content). Having two slots for grouping things that are used for similar things seemed like a good UX (making a few canonical things easy, like, "benchmark these alternative latent representations", "combine these labels into one", "fit a model on these conditions").
  b. One-dimensional annotations have a *simple* data type and a canonical storage format for grouping them (a DataFrame). But back in 2017, even this couldn't easily be stored persistently in an HDF5. So, when writing the on-disk format, I had to write a persistent serialization of a DataFrame and of multi-dimensional annotations. Both was feasible without too much work when done separately as each of these had simple data types, but I wouldn't have wanted to mix it (allowing for multi-dimensional annotations with mixed data types). @ivirshup implemented the latter years later.
6. Typo: https://github.com/scverse/anndata/pull/825/commits/24895a857a5335801146654115998a3a9825eca0
7. "h5ad": https://github.com/scverse/anndata/pull/825/commits/66962c58c43de10b09d1dbfc03c3a2aeee993482
8. "schema tracking": @ivirshup, next comment.

Debate:

1. "I think it is maybe inaccurate to imply AnnData is a data structure library (The AnnData object section).": @ivirshup & I discussed this remark in early August and thought we'd prefer to keep the term "data structure". While we agree that AnnData is based on existing data structures, it composes these to define a new data structure. We think this is a bit similar to how pandas stores data internally in numpy arrays, but composes them to give rise to a much richer structure, the DataFrame. Like pandas and other "data structure libraries", anndata offers APIs to conveniently interact with the new data structure.
2. "Clearer statement about filling a gap". We agree: https://github.com/scverse/anndata/pull/825/commits/4bbf466bbb14916a1714b3ac70465c690e9e2a39

Outlook:

1. "Lessons learned & benchmarks". @ivirshup will respond re benchmarking as he has spent much over the past years. I'm not sure whether there are very specific "lessons learned" for anndata. Clearly, today, a lot could be done differently. But back then, certain requirements made the adoption of say xarray hard to impossible. Isaac may have more things to say.
2. "Benchmarking": @ivirshup will address this.
3. "Self-comparative benchmarks": We agree!
4. "Low-level API": @ivirshup will address this.

Let us follow up with another post to address all outstanding remarks.
