### 2022-10-02 d6bd10f
Address Lance's JOSS comments (#825)
* Mentioned layers in main text

* Mention uns in main text

* Fix typo

* Clarify h5ad

* Clearer gap filling statement
