### 2022-10-09 bf9cafc
zarr partial reads
