### 2022-10-10 6659655
Add experimental support for `zstd` compression via hdf5plugin library
