+++ anndata/_core/anndata.py
@@ -15,6 +15,7 @@ from typing import Iterable, Sequence, Mapping, MutableMapping  # Generic ABCs
 from typing import Tuple, List  # Generic
 
 import h5py
+import hdf5plugin
 from natsort import natsorted
 import numpy as np
 from numpy import ma
@@ -1872,7 +1873,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     def write_h5ad(
         self,
         filename: Optional[PathLike] = None,
-        compression: Optional[Literal["gzip", "lzf"]] = None,
+        compression: Optional[Literal["gzip", "lzf", "zstd"]] = None,
         compression_opts: Union[int, Any] = None,
         force_dense: Optional[bool] = None,
         as_dense: Sequence[str] = (),
@@ -1897,9 +1898,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         filename
             Filename of data file. Defaults to backing file.
         compression
-            See the h5py :ref:`dataset_compression`.
+            For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
+            For [`zstd`], see the hdf5plugin :ref:`usage`. Experimental.
         compression_opts
-            See the h5py :ref:`dataset_compression`.
+            For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
+            For [`zstd`], see the hdf5plugin :ref:`usage`. Experimental.
+            Construct `compression_opts` using the helper functions provided
+            by the hdf5plugin package - for instance:
+            `hdf5plugin.Zstd(clevel=5).filter_options`
         as_dense
             Sparse arrays in AnnData object to write as dense. Currently only
             supports `X` and `raw/X`.
@@ -1913,6 +1919,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             raise ValueError("Provide a filename!")
         if filename is None:
             filename = self.filename
+        if compression in ["zstd"]:
+            compression = hdf5plugin.FILTERS[compression]
 
         _write_h5ad(
             Path(filename),
