### 2022-10-10 e861d41
Add .html to reference link for hdf5plugin docs
