### 2022-10-10 a3f8d9d
Fix link for hdf5plugin docs
