### 2022-10-11 64040e7
Fix hdf5plugin docs reference
Co-authored-by: Philipp A. <flying-sheep@web.de>
