### 2022-10-11 9110c2a
Fix hdf5plugin docs reference
Co-authored-by: Philipp A. <flying-sheep@web.de>
