### 2022-10-12 984be59
(feat): `read_dispatched` function
