### 2022-10-13 da03e4b
(feat): `zarr` read_dispatched
