### 2022-10-13 0eddd11
(fix): make `group` variable in function
