### 2022-10-13 dbac568
(feat): `read_dispatched` for `h5ad`
