+++ anndata/_io/zarr.py
@@ -80,9 +80,11 @@ def read_zarr(store: Union[str, Path, MutableMapping, zarr.Group]) -> AnnData:
         if k in {"obs", "var"}:
             return read_dataframe(group[k])
         return read_func(group[k])
-    
+
     def dispatch_anndata_args(group, args):
-        args["raw"] = _read_legacy_raw(group, args.get("raw"), read_dataframe, read_elem)
+        args["raw"] = _read_legacy_raw(
+            group, args.get("raw"), read_dataframe, read_elem
+        )
 
         if "X" in args:
             args["dtype"] = args["X"].dtype
@@ -90,11 +92,11 @@ def read_zarr(store: Union[str, Path, MutableMapping, zarr.Group]) -> AnnData:
         # Backwards compat to <0.7
         if isinstance(group["obs"], zarr.Array):
             _clean_uns(args)
-         
+
         return args
-    
-    
+
     return read_dispatched(f, dispatch_element, dispatch_anndata_args)
+
 
 @report_read_key_on_error
 def read_dataset(dataset: zarr.Array):
