### 2022-10-17 83612ba
add uns to_memory and remove old test
