### 2022-10-17 13a936e
refactor and added all tests except raw
