+++ anndata/_io/h5ad.py
@@ -189,7 +189,8 @@ def read_h5ad(
         changes to other slots like `obs` will not be written to disk in
         `backed` mode. If you would like save changes made to these slots
         of a `backed` :class:`~anndata.AnnData`, write them to a new file
-        (see :meth:`~anndata.AnnData.write`).
+        (see :meth:`~anndata.AnnData.write`). For an example, see
+        [here] (https://anndata-tutorials.readthedocs.io/en/latest/getting-started.html#Partial-reading-of-large-data).
     as_sparse
         If an array was saved as dense, passing its name here will read it as
         a sparse_matrix, by chunk of size `chunk_size`.
