### 2022-10-18 098f631
added raw to_memory tests
