### 2022-10-19 1cccc59
Re-add schema to README
