### 2022-10-19 211340b
use URL instead of path in README so PyPI can render it
