+++ anndata/_io/specs/__init__.py
@@ -1,2 +1,2 @@
 from . import methods
-from .registry import write_elem, read_elem, get_spec, _REGISTRY
+from .registry import write_elem, get_spec, read_elem, _REGISTRY, Reader
