+++ anndata/_io/specs/methods.py
@@ -76,7 +76,7 @@ H5Group = h5py.Group
 
 @_REGISTRY.register_read(H5Group, IOSpec("", ""))
 @_REGISTRY.register_read(H5Array, IOSpec("", ""))
-def read_basic(elem):
+def read_basic(elem, _reader):
     from anndata._io import h5ad
 
     warn(
@@ -89,14 +89,14 @@ def read_basic(elem):
         # Backwards compat sparse arrays
         if "h5sparse_format" in elem.attrs:
             return SparseDataset(elem).to_memory()
-        return {k: read_elem(v) for k, v in elem.items()}
+        return {k: _reader.read_elem(v) for k, v in elem.items()}
     elif isinstance(elem, h5py.Dataset):
         return h5ad.read_dataset(elem)  # TODO: Handle legacy
 
 
 @_REGISTRY.register_read(ZarrGroup, IOSpec("", ""))
 @_REGISTRY.register_read(ZarrArray, IOSpec("", ""))
-def read_basic_zarr(elem):
+def read_basic_zarr(elem, _reader):
     from anndata._io import zarr
 
     warn(
@@ -109,7 +109,7 @@ def read_basic_zarr(elem):
         # Backwards compat sparse arrays
         if "h5sparse_format" in elem.attrs:
             return SparseDataset(elem).to_memory()
-        return {k: read_elem(v) for k, v in elem.items()}
+        return {k: _reader.read_elem(v) for k, v in elem.items()}
     elif isinstance(elem, ZarrArray):
         return zarr.read_dataset(elem)  # TODO: Handle legacy
 
@@ -230,7 +230,7 @@ def write_anndata(f, k, adata, dataset_kwargs=MappingProxyType({})):
 @_REGISTRY.register_read(H5Group, IOSpec("raw", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("anndata", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("raw", "0.1.0"))
-def read_anndata(elem):
+def read_anndata(elem, _reader):
     d = {}
     for k in [
         "X",
@@ -245,7 +245,7 @@ def read_anndata(elem):
         "raw",
     ]:
         if k in elem:
-            d[k] = read_elem(elem[k])
+            d[k] = _reader.read_elem(elem[k])
         if "X" in d:
             d["dtype"] = d["X"].dtype
     return AnnData(**d)
@@ -267,8 +267,8 @@ def write_raw(f, k, raw, dataset_kwargs=MappingProxyType({})):
 
 @_REGISTRY.register_read(H5Group, IOSpec("dict", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("dict", "0.1.0"))
-def read_mapping(elem):
-    return {k: read_elem(v) for k, v in elem.items()}
+def read_mapping(elem, _reader):
+    return {k: _reader.read_elem(v) for k, v in elem.items()}
 
 
 @_REGISTRY.register_write(H5Group, OverloadedDict, IOSpec("dict", "0.1.0"))
@@ -319,7 +319,7 @@ def write_basic_dask(f, k, elem, dataset_kwargs=MappingProxyType({})):
 @_REGISTRY.register_read(H5Array, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("string-array", "0.2.0"))
-def read_array(elem):
+def read_array(elem, _reader):
     return elem[()]
 
 
[…]
