### 2022-10-21 scverse/anndata#769 @rcannood
@nlhepler Next time, please create a separate issue for your own comments. It's becoming hard to track what still needs to be done :)
