### 2022-10-21 scverse/anndata#769 @rcannood
To recap, I think my remaining issues are:

General issues
- [ ] **Repository** It seems the repository has been renamed from theislab/anndata to scverse/anndata. Please fix this in JOSS.

Functionality
- [ ] **Performance** "If there are any performance claims of the software, have they been confirmed?"
  Would it be possible to include an atlas-sized dataset (e.g. >100'000 cells) in the benchmarks included on the website / somewhere?

Documentation
- [ ] **Installation instructions** I don't see a "clearly-stated list of dependencies" in the documentation, as required by JOSS. I think the installation instructions could be expanded upon (e.g. https://scanpy.readthedocs.io/en/stable/installation.html).

Let's make one more push towards getting AnnData published in JOSS. Let's go @falexwolf @ivirshup !
