### 2022-10-31 a96205c
fix doc typos and outdated link (#837)
