### 2022-10-31 7c03acb
[DOCS] Explain backed is for X only (#812)
* [DOCS] Explain backed is for X only
* more explanation of backed in read_h5ad
* typo

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
