+++ docs/index.rst
@@ -35,6 +35,7 @@ Latest additions
    api
    concatenation
    fileformat-prose
+   interoperability
    benchmarks
    contributing
    release-notes/index
