### 2022-11-03 6585052
Update anndata/_core/index.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
