### 2022-11-03 e0e3da5
Update anndata/_core/index.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
