### 2022-11-03 c26ecce
Update anndata/_io/specs/methods.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
