+++ anndata/_io/specs/methods.py
@@ -10,7 +10,6 @@ from warnings import warn
 import h5py
 import numpy as np
 import pandas as pd
-
 from scipy import sparse
 
 import anndata as ad
