### 2022-11-03 7bcab52
Update anndata/_io/specs/methods.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
