### 2022-11-03 4d08070
Update anndata/tests/helpers.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
