### 2022-11-03 6c4dd44
Update anndata/tests/helpers.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
