### 2022-11-03 1f7b90a
merge two helper funcs
