+++ anndata/tests/test_views.py
@@ -19,7 +19,7 @@ from anndata.tests.helpers import (
     single_subset,
     assert_equal,
     as_dense_dask_array,
-    gen_adata_dask_args,
+    GEN_ADATA_DASK_ARGS,
 )
 
 # ------------------------------------------------------------------------------
@@ -337,7 +337,7 @@ def test_set_subset_varm(adata, subset_func):
 
 @pytest.mark.parametrize("attr", ["obsm", "varm", "obsp", "varp", "layers"])
 def test_view_failed_delitem(attr):
-    adata = gen_adata((10, 10), **gen_adata_dask_args)
+    adata = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
     view = adata[5:7, :][:, :5]
     adata_hash = joblib.hash(adata)
     view_hash = joblib.hash(view)
@@ -352,7 +352,7 @@ def test_view_failed_delitem(attr):
 
 @pytest.mark.parametrize("attr", ["obsm", "varm", "obsp", "varp", "layers"])
 def test_view_delitem(attr):
-    adata = gen_adata((10, 10), **gen_adata_dask_args)
+    adata = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
     getattr(adata, attr)["to_delete"] = np.ones((10, 10))
     # Shouldn’t be a subclass, should be an ndarray
     assert type(getattr(adata, attr)["to_delete"]) is np.ndarray
@@ -373,7 +373,7 @@ def test_view_delitem(attr):
     "attr", ["X", "obs", "var", "obsm", "varm", "obsp", "varp", "layers", "uns"]
 )
 def test_view_delattr(attr, subset_func):
-    base = gen_adata((10, 10), **gen_adata_dask_args)
+    base = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
     orig_hash = joblib.hash(base)
     subset = base[subset_func(base.obs_names), subset_func(base.var_names)]
     empty = ad.AnnData(obs=subset.obs[[]], var=subset.var[[]])
@@ -391,7 +391,7 @@ def test_view_delattr(attr, subset_func):
 )
 def test_view_setattr_machinery(attr, subset_func, subset_func2):
     # Tests that setting attributes on a view doesn't mess anything up too bad
-    adata = gen_adata((10, 10), **gen_adata_dask_args)
+    adata = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
     view = adata[subset_func(adata.obs_names), subset_func2(adata.var_names)]
 
     actual = view.copy()
@@ -462,7 +462,7 @@ def test_view_of_view_modification():
 
 
 def test_double_index(subset_func, subset_func2):
-    adata = gen_adata((10, 10), **gen_adata_dask_args)
+    adata = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
     obs_subset = subset_func(adata.obs_names)
     var_subset = subset_func2(adata.var_names)
     v1 = adata[obs_subset, var_subset]
