### 2022-11-03 8c0b592
Update anndata/tests/helpers.py
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
