### 2022-11-03 08cd3ef
array_typer_dense removed
