+++ anndata/tests/test_concatenate.py
@@ -72,26 +72,12 @@ def make_idx_tuple(idx, axis):
 
 
 # Will call func(sparse_matrix) so these types should be sparse compatible
-# See array_type_dense if only dense arrays are expected as input.
-@pytest.fixture(
-    params=[asarray, sparse.csr_matrix, sparse.csc_matrix],
-    ids=["np_array", "scipy_csr", "scipy_csc"],
-)
-def array_type(request):
-    return request.param
-
-
-# Give this array_type_dense instead of array_type if
-# the expected result is compared only content wise and
-# array types aren't expected to be the same
-# in other words, when:
-# test_assert(arr1,arr2) = test_assert(arr1.todense(),arr2.todense())
-# is the expected behaviour.
+# See array_type if only dense arrays are expected as input.
 @pytest.fixture(
     params=[asarray, sparse.csr_matrix, sparse.csc_matrix, as_dense_dask_array],
     ids=["np_array", "scipy_csr", "scipy_csc", "dask_array"],
 )
-def array_type_dense(request):
+def array_type(request):
     return request.param
 
 
@@ -165,10 +151,8 @@ def test_concat_interface_errors():
         (lambda x, **kwargs: x[0].concatenate(x[1:], **kwargs), True),
     ],
 )
-def test_concatenate_roundtrip(
-    join_type, array_type_dense, concat_func, backwards_compat
-):
-    adata = gen_adata((100, 10), X_type=array_type_dense, **GEN_ADATA_DASK_ARGS)
+def test_concatenate_roundtrip(join_type, array_type, concat_func, backwards_compat):
+    adata = gen_adata((100, 10), X_type=array_type, **GEN_ADATA_DASK_ARGS)
 
     remaining = adata.obs_names
     subsets = []
@@ -263,10 +247,10 @@ def test_concatenate_dense():
     assert np.allclose(var_ma.compressed(), var_ma_ref.compressed())
 
 
-def test_concatenate_layers(array_type_dense, join_type):
+def test_concatenate_layers(array_type, join_type):
     adatas = []
     for _ in range(5):
-        a = array_type_dense(sparse.random(100, 200, format="csr"))
+        a = array_type(sparse.random(100, 200, format="csr"))
         adatas.append(AnnData(X=a, layers={"a": a}))
 
     merged = adatas[0].concatenate(adatas[1:], join=join_type)
@@ -434,24 +418,11 @@ def test_concatenate_layers_misaligned(array_type, join_type):
     assert_equal(merged.X, merged.layers["a"])
 
 
-def test_concatenate_layers_misaligned_dense(array_type_dense, join_type):
-    adatas = []
-    for _ in range(5):
-        a = array_type_dense(np.random.rand(100, 200))
-        adata = AnnData(X=a, layers={"a": a}, dtype=a.dtype)
-        adatas.append(
-            adata[:, np.random.choice(adata.var_names, 150, replace=False)].copy()
-        )
-
-    merged = adatas[0].concatenate(adatas[1:], join=join_type)
-    assert_equal(merged.X, merged.layers["a"])
-
-
-def test_concatenate_layers_outer(array_type_dense, fill_val):
+def test_concatenate_layers_outer(array_type, fill_val):
     # Testing that issue #368 is fixed
     a = AnnData(
         X=np.ones((10, 20)),
-        layers={"a": array_type_dense(sparse.random(10, 20, format="csr"))},
+        layers={"a": array_type(sparse.random(10, 20, format="csr"))},
[…]
