### 2022-11-03 e39c508
to_memory is now in-place
