### 2022-11-03 c2629c4
to_memory has copy option
