+++ anndata/tests/test_dask.py
@@ -135,6 +135,44 @@ def test_dask_to_memory_check_array_types(adata, tmp_path, diskfmt):
     assert isinstance(orig.varm["a"], np.ndarray)
 
 
+def test_dask_to_memory_copy_check_array_types(adata, tmp_path, diskfmt):
+    import dask.array as da
+    import numpy as np
+
+    pth = tmp_path / f"test_write.{diskfmt}"
+    write = lambda x, y: getattr(x, f"write_{diskfmt}")(y)
+    read = lambda x: getattr(ad, f"read_{diskfmt}")(x)
+
+    M, N = adata.X.shape
+    adata.obsm["a"] = da.random.random((M, 10))
+    adata.obsm["b"] = da.random.random((M, 10))
+    adata.varm["a"] = da.random.random((N, 10))
+
+    orig = adata
+    write(orig, pth)
+    curr = read(pth)
+
+    mem = orig.to_memory(copy=True)
+
+    with pytest.raises(Exception):
+        assert_equal(curr.obsm["a"], curr.obsm["b"])
+
+    assert_equal(curr.varm["a"], orig.varm["a"])
+    assert_equal(curr.obsm["a"], orig.obsm["a"])
+    assert_equal(mem.obsm["a"], orig.obsm["a"])
+    assert_equal(mem.varm["a"], orig.varm["a"])
+
+    assert isinstance(curr.X, np.ndarray)
+    assert isinstance(curr.obsm["a"], np.ndarray)
+    assert isinstance(curr.varm["a"], np.ndarray)
+    assert isinstance(mem.X, np.ndarray)
+    assert isinstance(mem.obsm["a"], np.ndarray)
+    assert isinstance(mem.varm["a"], np.ndarray)
+    assert isinstance(orig.X, DaskArray)
+    assert isinstance(orig.obsm["a"], DaskArray)
+    assert isinstance(orig.varm["a"], DaskArray)
+
+
 def test_dask_copy_check_array_types(adata):
     import dask.array as da
 
@@ -190,6 +228,7 @@ def test_dask_to_memory_unbacked():
     assert isinstance(orig.uns["da"]["da"], DaskArray)
 
     curr = orig.to_memory()
+
     assert_equal(orig, curr)
     assert isinstance(curr.X, np.ndarray)
     assert isinstance(curr.obsm["da"], np.ndarray)
@@ -201,6 +240,28 @@ def test_dask_to_memory_unbacked():
     assert isinstance(orig.layers["da"], np.ndarray)
     assert isinstance(orig.varm["da"], np.ndarray)
     assert isinstance(orig.uns["da"]["da"], np.ndarray)
+
+
+# Test if dask arrays turn into numpy arrays after to_memory is called
+def test_dask_to_memory_copy_unbacked():
+    import numpy as np
+
+    orig = gen_adata((15, 10), X_type=as_dense_dask_array, **GEN_ADATA_DASK_ARGS)
+    orig.uns = {"da": {"da": as_dense_dask_array(np.ones(12))}}
+
+    curr = orig.to_memory(copy=True)
+
+    assert_equal(orig, curr)
+    assert isinstance(curr.X, np.ndarray)
+    assert isinstance(curr.obsm["da"], np.ndarray)
+    assert isinstance(curr.varm["da"], np.ndarray)
+    assert isinstance(curr.layers["da"], np.ndarray)
+    assert isinstance(curr.uns["da"]["da"], np.ndarray)
+    assert isinstance(orig.X, DaskArray)
+    assert isinstance(orig.obsm["da"], DaskArray)
+    assert isinstance(orig.layers["da"], DaskArray)
[…]
