### 2022-11-03 d141dba
refactor to_memory code
