+++ docs/release-notes/0.7.6.md
@@ -1,0 +1,24 @@
+# 0.7.6 {small}`11 April, 2021`
+
+```{rubric} New features
+```
+
+- Added {meth}`anndata.AnnData.to_memory` for returning an in memory object from a backed one {pr}`470` {pr}`542` {smaller}`V Bergen` {smaller}`I Virshup`
+- {meth}`anndata.AnnData.write_loom` now writes `obs_names` and `var_names` using the `Index`'s `.name` attribute, if set {pr}`538` {smaller}`I Virshup`
+
+```{rubric} Bug fixes
+```
+
+- Fixed bug where `np.str_` column names errored at write time {pr}`457` {smaller}`I Virshup`
+- Fixed "value.index does not match parent’s axis 0/1 names" error triggered when a data frame is stored in obsm/varm after obs_names/var_names is updated {pr}`461` {smaller}`G Eraslan`
+- Fixed `adata.write_csvs` when `adata` is a view {pr}`462` {smaller}`I Virshup`
+- Fixed null values being converted to strings when strings are converted to categorical {pr}`529` {smaller}`I Virshup`
+- Fixed handling of compression key word arguments {pr}`536` {smaller}`I Virshup`
+- Fixed copying a backed `AnnData` from changing which file the original object points at {pr}`533` {smaller}`ilia-kats`
+- Fixed a bug where calling `AnnData.concatenate` an `AnnData` with no variables would error {pr}`537` {smaller}`I Virshup`
+
+```{rubric} Deprecations
+```
+
+- Passing positional arguments to {func}`anndata.read_loom` besides the path is now deprecated {pr}`538` {smaller}`I Virshup`
+- {func}`anndata.read_loom` arguments `obsm_names` and `varm_names` are now deprecated in favour of `obsm_mapping` and `varm_mapping` {pr}`538` {smaller}`I Virshup`
