+++ docs/tutorials.md
@@ -1,0 +1,9 @@
+# Tutorials
+
+For a quick introduction to `AnnData`, check out {tutorial}`getting-started`.
+
+For working with the experimental data loaders (see {ref}`experimental_api`):
+
+- For an example of working with pytorch {tutorial}`annloader`
+- For lazily concatenating multiple files {tutorial}`anncollection`
+- For interfacing pytorch models with lazy collections {tutorial}`anncollection-annloader`
