+++ docs/release-notes/0.6.0.md
@@ -1,0 +1,33 @@
+# 0.6.\* {small}`2019-*-*`
+
+- better support for aligned mappings (obsm, varm, layers)
+  {noteversion}`0.6.22` {pr}`155` {smaller}`I Virshup`
+- convenience accesors {func}`~anndata.AnnData.obs_vector`, {func}`~anndata.AnnData.var_vector` for 1d arrays.
+  {noteversion}`0.6.21` {pr}`144` {smaller}`I Virshup`
+- compatibility with Scipy >=1.3 by removing `IndexMixin` dependency.
+  {noteversion}`0.6.20` {pr}`151` {smaller}`P Angerer`
+- bug fix for second-indexing into views.
+  {noteversion}`0.6.19` {smaller}`P Angerer`
+- bug fix for reading excel files.
+  {noteversion}`0.6.19` {smaller}`A Wolf`
+- changed default compression to `None` in {func}`~anndata.AnnData.write_h5ad` to speed up read and write, disk space use is usually less critical.
+  {noteversion}`0.6.16` {smaller}`A Wolf`
+- maintain dtype upon copy.
+  {noteversion}`0.6.13` {smaller}`A Wolf`
+- {attr}`~anndata.AnnData.layers` inspired by [.loom](https://loompy.org) files allows their information lossless reading via {func}`~anndata.read_loom`.
+  {noteversion}`0.6.7`–{noteversion}`0.6.9` {pr}`46` & {pr}`48` {smaller}`S Rybakov`
+- support for reading zarr files: {func}`~anndata.read_zarr`
+  {noteversion}`0.6.7` {pr}`38` {smaller}`T White`
+- initialization from pandas DataFrames
+  {noteversion}`0.6.` {smaller}`A Wolf`
+- iteration over chunks {func}`~anndata.AnnData.chunked_X` and {func}`~anndata.AnnData.chunk_X`
+  {noteversion}`0.6.1` {pr}`20` {smaller}`S Rybakov`
+
+# 0.6.0 {small}`1 May, 2018`
+
+- compatibility with Seurat converter
+- tremendous speedup for {func}`~anndata.AnnData.concatenate`
+- bug fix for deep copy of unstructured annotation after slicing
+- bug fix for reading HDF5 stored single-category annotations
+- `'outer join'` concatenation: adds zeros for concatenation of sparse data and nans for dense data
+- better memory efficiency in loom exports
