+++ docs/_key_contributors.md
@@ -1,0 +1,10 @@
+:::{sidebar} Key Contributors
+- [Isaac Virshup]: anndata >= 0.7, diverse contributions
+- Sergei Rybakov: diverse contributions
+- [Alex Wolf]: initial conception/development
+- Philipp Angerer: initial conception/development, software quality
+:::
+
+[alex wolf]: https://twitter.com/falexwolf
+[contributions graph]: https://github.com/scverse/anndata/graphs/contributors
+[isaac virshup]: https://twitter.com/ivirshup
