+++ docs/release-notes/0.8.1.md
@@ -1,4 +1,4 @@
-# 0.8.1 {small}`the future`
+### 0.8.1 {small}`the future`
 
 ```{rubric} Bug fixes
 ```
