+++ docs/release-notes/0.7.6.md
@@ -1,4 +1,4 @@
-# 0.7.6 {small}`11 April, 2021`
+### 0.7.6 {small}`11 April, 2021`
 
 ```{rubric} New features
 ```
