+++ docs/release-notes/0.5.0.md
@@ -1,4 +1,4 @@
-# 0.5.0 {small}`9 February, 2018`
+### 0.5.0 {small}`9 February, 2018`
 
 - inform about duplicates in {class}`~anndata.AnnData.var_names` and resolve them using {func}`~anndata.AnnData.var_names_make_unique`
 - automatically remove unused categories after slicing
