+++ docs/contributing.md
@@ -2,6 +2,5 @@
 
 AnnData follows the development practices outlined in the [Scanpy contribution guide](https://scanpy.readthedocs.io/en/latest/dev/index.html).
 
-```{eval-rst}
-.. include:: _key_contributors.rst
+```{include} _key_contributors.md
 ```
