+++ docs/release-notes/0.6.0.md
@@ -1,4 +1,4 @@
-# 0.6.\* {small}`2019-*-*`
+### 0.6.\* {small}`2019-*-*`
 
 - better support for aligned mappings (obsm, varm, layers)
   {noteversion}`0.6.22` {pr}`155` {smaller}`I Virshup`
@@ -23,7 +23,7 @@
 - iteration over chunks {func}`~anndata.AnnData.chunked_X` and {func}`~anndata.AnnData.chunk_X`
   {noteversion}`0.6.1` {pr}`20` {smaller}`S Rybakov`
 
-# 0.6.0 {small}`1 May, 2018`
+### 0.6.0 {small}`1 May, 2018`
 
 - compatibility with Seurat converter
 - tremendous speedup for {func}`~anndata.AnnData.concatenate`
