+++ docs/release-notes/0.7.4.md
@@ -1,4 +1,4 @@
-# 0.7.4 {small}`10 July, 2020`
+### 0.7.4 {small}`10 July, 2020`
 
 ~~~{rubric} Concatenation overhaul {pr}`378` {smaller}`I Virshup`
 ~~~
