+++ docs/release-notes/0.7.2.md
@@ -1,4 +1,4 @@
-# 0.7.2 {small}`15 May, 2020`
+### 0.7.2 {small}`15 May, 2020`
 
 ~~~{rubric} Concatenation overhaul {smaller}`I Virshup`
 ~~~
