+++ docs/release-notes/0.7.5.rst
@@ -1,12 +1,0 @@
-0.7.5 :small:`12 November, 2020`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. rubric:: Functionality
-
-- Added ipython tab completion and a useful return from `.keys` to `adata.uns` :pr:`415` :smaller:`I Virshup`
-
-.. rubric:: Bug fixes
-
-- Compatibility with `h5py>=3` strings :pr:`444` :smaller:`I Virshup`
-- Allow `adata.raw = None`, as is documented :pr:`447` :smaller:`I Virshup`
-- Fix warnings from pandas 1.1 :pr:`425` :smaller:`I Virshup`
