### 2022-11-04 5881238
Move to markdown readme, remove markdown templates
