+++ docs/_templates/autosummary/base.md
@@ -1,7 +1,0 @@
----
-github_url: '{{ fullname | github_url }}'
----
-
-{% extends "!autosummary/base.rst" %}
-
-% http://www.sphinx-doc.org/en/stable/ext/autosummary.html#customizing-templates
