+++ docs/_templates/autosummary/class.md
@@ -1,41 +1,0 @@
----
-github_url: '{{ fullname | github_url }}'
----
-
-{{ fullname | escape | underline}}
-
-```{eval-rst}
-.. currentmodule:: {{ module }}
-```
-
-% add toctree option to make autodoc generate the pages
-
-```{eval-rst}
-.. autoclass:: {{ objname }}
-
-   {% block attributes %}
-   {% if attributes %}
-   .. rubric:: Attributes
-
-   .. autosummary::
-      :toctree: .
-   {% for item in attributes %}
-      ~{{ fullname }}.{{ item }}
-   {%- endfor %}
-   {% endif %}
-   {% endblock %}
-
-   {% block methods %}
-   {% if methods %}
-   .. rubric:: Methods
-
-   .. autosummary::
-      :toctree: .
-   {% for item in methods %}
-      {%- if item != '__init__' %}
-      ~{{ fullname }}.{{ item }}
-      {%- endif -%}
-   {%- endfor %}
-   {% endif %}
-   {% endblock %}
-```
