### 2022-11-04 a56bc21
Move concatenation docs back to rst, myst doesn't do doctests
