+++ docs/concatenation.rst
@@ -1,0 +1,288 @@
+Concatenation
+=============
+
+With :func:`~anndata.concat`, :class:`~anndata.AnnData` objects can be combined via a composition of two operations: concatenation and merging.
+
+* Concatenation is when we keep all sub elements of each object, and stack these elements in an ordered way.
+* Merging is combining a set of collections into one resulting collection which contains elements from the objects.
+
+.. note::
+
+    This function borrows from similar functions in pandas_ and xarray_. Argument which are used to control concatenation are modeled after :func:`pandas.concat` while strategies for merging are inspired by :func:`xarray.merge`'s `compat` argument.
+
+.. _pandas: https://pandas.pydata.org
+.. _xarray: http://xarray.pydata.org
+
+Concatenation
+-------------
+
+Let's start off with an example:
+
+    >>> import scanpy as sc, anndata as ad, numpy as np, pandas as pd
+    >>> from scipy import sparse
+    >>> from anndata import AnnData
+    >>> pbmc = sc.datasets.pbmc68k_reduced()
+    >>> pbmc
+    AnnData object with n_obs × n_vars = 700 × 765
+        obs: 'bulk_labels', 'n_genes', 'percent_mito', 'n_counts', 'S_score', 'G2M_score', 'phase', 'louvain'
+        var: 'n_counts', 'means', 'dispersions', 'dispersions_norm', 'highly_variable'
+        uns: 'bulk_labels_colors', 'louvain', 'louvain_colors', 'neighbors', 'pca', 'rank_genes_groups'
+        obsm: 'X_pca', 'X_umap'
+        varm: 'PCs'
+        obsp: 'distances', 'connectivities'
+
+If we split this object up by clusters of observations, then stack those subsets we'll obtain the same values – just ordered differently.
+
+    >>> groups = pbmc.obs.groupby("louvain").indices
+    >>> pbmc_concat = ad.concat([pbmc[inds] for inds in groups.values()], merge="same")
+    >>> assert np.array_equal(pbmc.X, pbmc_concat[pbmc.obs_names].X)
+    >>> pbmc_concat
+    AnnData object with n_obs × n_vars = 700 × 765
+        obs: 'bulk_labels', 'n_genes', 'percent_mito', 'n_counts', 'S_score', 'G2M_score', 'phase', 'louvain'
+        var: 'n_counts', 'means', 'dispersions', 'dispersions_norm', 'highly_variable'
+        obsm: 'X_pca', 'X_umap'
+        varm: 'PCs'
+
+Note that we concatenated along the observations by default, and that most elements aligned to the observations were concatenated as well.
+A notable exception is :attr:`~anndata.AnnData.obsp`, which can be re-enabled with the `pairwise` keyword argument.
+This is because it's not obvious that combining graphs or distance matrices padded with 0s is particularly useful, and may be unintuitive.
+
+Inner and outer joins
+~~~~~~~~~~~~~~~~~~~~~
+
+When the variables present in the objects to be concatenated aren't exactly the same, you can choose to take either the intersection or union of these variables.
+This is otherwise called taking the `"inner"` (intersection) or `"outer"` (union) join.
+For example, given two anndata objects with differing variables:
+
+    >>> a = AnnData(sparse.eye(3), var=pd.DataFrame(index=list("abc")))
+    >>> b = AnnData(sparse.eye(2), var=pd.DataFrame(index=list("ba")))
+    >>> ad.concat([a, b], join="inner").X.toarray()
+    array([[1., 0.],
+           [0., 1.],
+           [0., 0.],
+           [0., 1.],
+           [1., 0.]], dtype=float32)
[…]
