### 2022-11-04 1654b25
to_mempry works as expected
