+++ anndata/tests/test_io_partial.py
@@ -2,9 +2,9 @@ from anndata import AnnData
 from anndata._io.specs.registry import read_elem_partial
 from anndata._io import write_zarr
 from scipy.sparse import csr_matrix
+from pathlib import Path
 import numpy as np
 import pytest
-from pathlib import Path
 import zarr
 
 X = np.array([[1.0, 0.0, 3.0], [4.0, 0.0, 6.0], [0.0, 8.0, 0.0]], dtype="float32")
@@ -15,8 +15,6 @@ X_check = np.array([[4.0, 0.0], [0.0, 8.0]], dtype="float32")
 def test_zarr_read_partial(tmp_path, typ):
 
     adata = AnnData(X=typ(X))
-
-    print(tmp_path)
 
     path = Path(tmp_path) / "test_zarr.zarr"
 
