### 2022-11-06 9e9441d
Fix zarr partial reads for sparse and dense arrays (#827)
* zarr partial reads

* add tests

* rm print
