### 2022-11-08 d662885
Consolidate to_memory for backed and unbacked cases
