+++ docs/release-notes/0.9.0.rst
@@ -5,6 +5,7 @@
 
 * Unordered categorical columns are no longer cast to object during :func:`anndata.concat` :pr:`763` :smaller:`ivirshup`
 * Expanded docstring more documentation for ``backed`` argument of :func:`anndata.read_h5ad` :pr:`812` :smaller:`jeskowagner`
+* Added support for dask arrays :pr:`813` :smaller:`syelman` :smaller:`rahulbshrestha`
 
 .. rubric:: Bug fixes
 
