### 2022-11-08 bc9f886
Dask array support (#813)
* copy paste util func

* some init commit

* add some tests to see where it fails

* add new ignored filenames

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update anndata/tests/test_dask.py

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>

* typo

* add da.Array to helpers.gen_adata

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* minor fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

[…]
