+++ anndata/tests/test_inplace_subset.py
@@ -2,13 +2,19 @@ import numpy as np
 import pytest
 from scipy import sparse
 
-from anndata.tests.helpers import assert_equal, gen_adata, subset_func
+from anndata.tests.helpers import (
+    assert_equal,
+    gen_adata,
+    subset_func,
+    as_dense_dask_array,
+)
 from anndata.utils import asarray
+from anndata.compat import DaskArray
 
 
 @pytest.fixture(
-    params=[np.array, sparse.csr_matrix, sparse.csc_matrix],
-    ids=["np_array", "scipy_csr", "scipy_csc"],
+    params=[np.array, sparse.csr_matrix, sparse.csc_matrix, as_dense_dask_array],
+    ids=["np_array", "scipy_csr", "scipy_csc", "dask_array"],
 )
 def matrix_type(request):
     return request.param
