### 2022-11-08 3e1cad3
Handle non-csr/csc sparse matrices in anndata.write
Attempting to save an AnnData object with non-CSR/CSC sparse matrices throws an error:

```
File "/usr/src/singlecellopenproblems/openproblems/api/utils.py", line 58, in write_h5ad
      adata.write_h5ad(filename)
    File "/usr/local/lib/python3.8/site-packages/anndata/_core/anndata.py", line 1918, in write_h5ad
      _write_h5ad(
    File "/usr/local/lib/python3.8/site-packages/anndata/_io/h5ad.py", line 85, in write_h5ad
      write_elem(f, "X", adata.X, dataset_kwargs=dataset_kwargs)
    File "/usr/local/lib/python3.8/site-packages/anndat
[…]
