### 2022-11-08 a5e0311
Fix sparse_to_dense
