### 2022-11-08 c446a23
Test coo_matrix
