### 2022-11-15 cf7d07c
(feat): `write_dispatched`
