### 2022-11-15 1df6eb0
(feat): `write_h5ad` via `write_dispatched`
