+++ anndata/_io/__init__.py
@@ -8,7 +8,7 @@ from .read import (
     read_text,
     read_dispatched,
 )
-from .zarr import read_zarr
-from .h5ad import read_h5ad
-from .write import write_csvs, write_loom, _write_h5ad, write_zarr
+from .zarr import read_zarr, write_zarr
+from .h5ad import read_h5ad, write_h5ad
+from .write import write_csvs, write_loom
 from . import h5ad
