### 2022-11-18 0a50ebe
Revert "Disable dry-run"
This reverts commit 0697e8247abc2f0037a48844ffe4721c58894c2a.
