### 2022-11-21 a577623
(feat): `write_zarr` via `write_dispatched`
