+++ anndata/_io/specs/methods.py
@@ -212,18 +212,18 @@ def _read_partial(group, *, items=None, indices=(slice(None), slice(None))):
 
 @_REGISTRY.register_write(ZarrGroup, AnnData, IOSpec("anndata", "0.1.0"))
 @_REGISTRY.register_write(H5Group, AnnData, IOSpec("anndata", "0.1.0"))
-def write_anndata(f, k, adata, dataset_kwargs=MappingProxyType({})):
+def write_anndata(f, k, adata, _writer, dataset_kwargs=MappingProxyType({})):
     g = f.require_group(k)
-    write_elem(g, "X", adata.X, dataset_kwargs=dataset_kwargs)
-    write_elem(g, "obs", adata.obs, dataset_kwargs=dataset_kwargs)
-    write_elem(g, "var", adata.var, dataset_kwargs=dataset_kwargs)
-    write_elem(g, "obsm", dict(adata.obsm), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "varm", dict(adata.varm), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "obsp", dict(adata.obsp), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "varp", dict(adata.varp), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "layers", dict(adata.layers), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "uns", dict(adata.uns), dataset_kwargs=dataset_kwargs)
-    write_elem(g, "raw", adata.raw, dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "X", adata.X, dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "obs", adata.obs, dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "var", adata.var, dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "obsm", dict(adata.obsm), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "varm", dict(adata.varm), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "obsp", dict(adata.obsp), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "varp", dict(adata.varp), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "layers", dict(adata.layers), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "uns", dict(adata.uns), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "raw", adata.raw, dataset_kwargs=dataset_kwargs)
 
 
 @_REGISTRY.register_read(H5Group, IOSpec("anndata", "0.1.0"))
@@ -253,11 +253,11 @@ def read_anndata(elem, _reader):
 
 @_REGISTRY.register_write(H5Group, Raw, IOSpec("raw", "0.1.0"))
 @_REGISTRY.register_write(ZarrGroup, Raw, IOSpec("raw", "0.1.0"))
-def write_raw(f, k, raw, dataset_kwargs=MappingProxyType({})):
+def write_raw(f, k, raw, _writer, dataset_kwargs=MappingProxyType({})):
     g = f.create_group(k)
-    write_elem(g, "X", raw.X, dataset_kwargs=dataset_kwargs)
-    write_elem(g, "var", raw.var, dataset_kwargs=dataset_kwargs)
-    write_elem(g, "varm", dict(raw.varm), dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "X", raw.X, dataset_kwargs=dataset_kwargs)
+    _writer.write_elem(g, "var", raw.var, dataset_kwargs=dataset_kwargs)
[…]
