+++ anndata/experimental/__init__.py
@@ -1,3 +1,5 @@
+from types import MappingProxyType
+
 from .multi_files import AnnCollection
 from .pytorch import AnnLoader
 
@@ -9,6 +11,7 @@ __all__ = [
     "read_elem",
     "write_elem",
     "read_dispatched",
+    "write_dispatched",
 ]
 
 
@@ -18,3 +21,11 @@ def read_dispatched(store, callback):
     reader = Reader(_REGISTRY, callback=callback)
 
     return reader.read_elem(store)
+
+
+def write_dispatched(store, key, elem, callback, dataset_kwargs=MappingProxyType({})):
+    from anndata._io.specs import Writer, _REGISTRY
+
+    writer = Writer(_REGISTRY, callback=callback)
+
+    writer.write_elem(store, key, elem, dataset_kwargs=dataset_kwargs)

[…]
