### 2022-11-21 9158b57
Add documentation for hdf5plugin in write_h5ad. Remove hdf5plugin code. Remove hdf5plugin tests.
