+++ anndata/tests/test_readwrite.py
@@ -6,7 +6,6 @@ import tempfile
 import warnings
 
 import h5py
-import hdf5plugin
 import numpy as np
 import pandas as pd
 from pandas.api.types import is_categorical_dtype
@@ -276,7 +275,7 @@ def test_readwrite_equivalent_h5ad_zarr(typ):
 
 @pytest.mark.parametrize(
     "compression,compression_opts",
-    [(None, None), ("lzf", None), ("gzip", None), ("gzip", 8), ("zstd", None)],
+    [(None, None), ("lzf", None), ("gzip", None), ("gzip", 8)],
 )
 def test_hdf5_compression_opts(tmp_path, compression, compression_opts):
     # https://github.com/scverse/anndata/issues/497
@@ -294,11 +293,7 @@ def test_hdf5_compression_opts(tmp_path, compression, compression_opts):
     def check_compressed(key, value):
         if isinstance(value, h5py.Dataset) and value.shape != ():
             if compression is not None and value.compression != compression:
-                # TODO: determine why using hdf5plugin compression filters
-                # like `zstd` causes `value.compression == None`, despite
-                # still compressing/decrompressing data as expected.
-                if compression not in ["zstd"]:
-                    not_compressed.append(key)
+                not_compressed.append(key)
             elif (
                 compression_opts is not None
                 and value.compression_opts != compression_opts
