### 2022-11-21 e049bde
Add read as dask array example test
