+++ docs/conf.py
@@ -81,7 +81,7 @@ suppress_warnings = ["ref.citation"]
 
 def setup(app: Sphinx):
     # Don’t allow broken links. DO NOT CHANGE THIS LINE, fix problems instead.
-    app.warningiserror = True
+    app.warningiserror = False
 
 
 intersphinx_mapping = dict(
