+++ docs/tutorials/index.rst
@@ -1,6 +1,8 @@
 Tutorials
 =========
 
+.. TODO Figure out what text I want here, or at least fix links to tutorials
+
 For a quick introduction to `AnnData`, check out :tutorial:`getting-started`.
 
 For working with the experimental data loaders (see :ref:`experimental_api`):
@@ -8,3 +10,11 @@ For working with the experimental data loaders (see :ref:`experimental_api`):
 * For an example of working with pytorch :tutorial:`annloader`
 * For lazily concatenating multiple files :tutorial:`anncollection`
 * For interfacing pytorch models with lazy collections :tutorial:`anncollection-annloader`
+
+.. toctree::
+   :maxdepth: 4
+
+   notebooks/getting-started
+   notebooks/annloader
+   notebooks/anncollection
+   notebooks/anncollection-annloader
