### 2022-11-22 c422646
Docs formatting and removal of hdf5plugin from dependency list
