+++ anndata/_core/anndata.py
@@ -1931,14 +1931,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             Alternative compression filters such as `zstd` can be used by
             passing a compressor from the hdf5plugin library; see
             :ref:hdf5plugin:`usage` for more information. Experimental.
-            Usage example:
-            ```
-            import hdf5plugin
-            adata.write_h5ad(
-                filename,
-                compression=hdf5plugin.FILTERS["zstd"]
-            )
-            ```
+            Usage example::
+
+                import hdf5plugin
+                adata.write_h5ad(
+                    filename,
+                    compression=hdf5plugin.FILTERS["zstd"]
+                )
+            
             .. note::
                 Datasets written with hdf5plugin-provided compressors
                 cannot be opened without first loading the hdf5plugin
@@ -1946,21 +1946,22 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 compression filters such as `zstd`, consider writing to
                 `zarr` format instead of `h5ad`, as the `zarr` library
                 provides a more transparent compression pipeline. 
+
         compression_opts
             For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
             For alternative compressor filters provided by hdf5plugin,
             `compression_opts` can be constructed using the instantiated
              compressor's `filter_options` property; see :ref:hdf5plugin:`usage`
             for more information. Experimental.
-            Usage example (setting `zstd` compression level to 5):
-            ```
-            import hdf5plugin
-            adata.write_h5ad(
-                filename, 
-                compression=hdf5plugin.FILTERS["zstd"],
-                compression_opts=hdf5plugin.Zstd(clevel=5).filter_options
-            )
-            ```
+            Usage example (setting `zstd` compression level to 5)::
+            
+                import hdf5plugin
+                adata.write_h5ad(
+                    filename, 
+                    compression=hdf5plugin.FILTERS["zstd"],
+                    compression_opts=hdf5plugin.Zstd(clevel=5).filter_options
+                )
+            
         as_dense
             Sparse arrays in AnnData object to write as dense. Currently only
             supports `X` and `raw/X`.
