### 2022-11-22 baa265c
Typing for writer signature
