+++ anndata/__init__.py
@@ -17,6 +17,8 @@ if not within_flit():
         read_text,
         read_mtx,
         read_zarr,
+        read_dispatched,
+        write_dispatched
     )
     from ._warnings import OldFormatWarning, WriteWarning, ImplicitModificationWarning
 
