+++ anndata/_io/__init__.py
@@ -10,6 +10,6 @@ from .read import (
 )
 from .zarr import read_zarr, write_zarr
 from .h5ad import read_h5ad, write_h5ad
-from .write import write_csvs, write_loom
+from .write import write_csvs, write_loom, write_dispatched
 from . import h5ad
 from . import zarr
