+++ anndata/_core/anndata.py
@@ -166,8 +166,6 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         If passing a :class:`~numpy.ndarray`, it needs to have a structured datatype.
     layers
         Key-indexed multi-dimensional arrays aligned to dimensions of `X`.
-    dtype
-        Data type used for storage.
     shape
         Shape tuple (#observations, #variables). Can only be provided if `X` is `None`.
     filename
@@ -451,27 +449,21 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 raise ValueError("`shape` needs to be `None` if `X` is not `None`.")
             _check_2d_shape(X)
             # if type doesn’t match, a copy is made, otherwise, use a view
-            if dtype is None and X.dtype != np.float32:
+            if dtype is not None:
+                # TODO: Decide on whether this should warn yet
                 warnings.warn(
-                    f"X.dtype being converted to np.float32 from {X.dtype}. In the next "
-                    "version of anndata (0.9) conversion will not be automatic. Pass "
-                    "dtype explicitly to avoid this warning. Pass "
-                    "`AnnData(X, dtype=X.dtype, ...)` to get the future behavour.",
-                    FutureWarning,
-                    stacklevel=3,
+                    "The dtype argument is deprecated and will be removed in 0.10.0",
+                    DeprecationWarning,
                 )
-                dtype = np.float32
-            elif dtype is None:
-                dtype = np.float32
-            if issparse(X) or isinstance(X, ma.MaskedArray):
-                # TODO: maybe use view on data attribute of sparse matrix
-                #       as in readwrite.read_10x_h5
-                if X.dtype != np.dtype(dtype):
+                if issparse(X) or isinstance(X, ma.MaskedArray):
+                    # TODO: maybe use view on data attribute of sparse matrix
+                    #       as in readwrite.read_10x_h5
+                    if X.dtype != np.dtype(dtype):
+                        X = X.astype(dtype)
+                elif isinstance(X, (ZarrArray, DaskArray)):
                     X = X.astype(dtype)
-            elif isinstance(X, (ZarrArray, DaskArray)):
-                X = X.astype(dtype)
-            else:  # is np.ndarray or a subclass, convert to true np.ndarray
-                X = np.array(X, dtype, copy=False)
+                else:  # is np.ndarray or a subclass, convert to true np.ndarray
+                    X = np.array(X, dtype, copy=False)
             # data matrix and shape
             self._X = X
             self._n_obs, self._n_vars = self._X.shape
@@ -1243,11 +1235,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         Same as `adata = adata[:, index]`, but inplace.
         """
         adata_subset = self[:, index].copy()
-        if adata_subset._has_X():
-            dtype = adata_subset.X.dtype
-        else:
-            dtype = None
-        self._init_as_actual(adata_subset, dtype=dtype)
+        self._init_as_actual(adata_subset)
 
     def _inplace_subset_obs(self, index: Index1D):
         """\
@@ -1256,11 +1244,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         Same as `adata = adata[index, :]`, but inplace.
         """
         adata_subset = self[index].copy()
-        if adata_subset._has_X():
-            dtype = adata_subset.X.dtype
-        else:
-            dtype = None
-        self._init_as_actual(adata_subset, dtype=dtype)
[…]
