+++ anndata/_io/h5ad.py
@@ -146,16 +146,6 @@ def read_h5ad_backed(filename: Union[str, Path], mode: Literal["r", "r+"]) -> An
 
     d["raw"] = _read_raw(f, attrs={"var", "varm"})
 
-    X_dset = f.get("X", None)
-    if X_dset is None:
-        pass
-    elif isinstance(X_dset, h5py.Group):
-        d["dtype"] = X_dset["data"].dtype
-    elif hasattr(X_dset, "dtype"):
-        d["dtype"] = f["X"].dtype
-    else:
-        raise ValueError()
-
     # Backwards compat to <0.7
     if isinstance(f["obs"], h5py.Dataset):
         _clean_uns(d)
@@ -246,16 +236,6 @@ def read_h5ad(
                 d[k] = read_elem(f[k])
 
         d["raw"] = _read_raw(f, as_sparse, rdasp)
-
-        X_dset = f.get("X", None)
-        if X_dset is None:
-            pass
-        elif isinstance(X_dset, h5py.Group):
-            d["dtype"] = X_dset["data"].dtype
-        elif hasattr(X_dset, "dtype"):
-            d["dtype"] = f["X"].dtype
-        else:
-            raise ValueError()
 
         # Backwards compat to <0.7
         if isinstance(f["obs"], h5py.Dataset):
