### 2022-11-22 fabe5dd
Release note + test fix
