### 2022-11-22 d731544
Make deprecation of dtype argument pending
