### 2022-11-22 730be92
Stop using dtype argument (#854)
* Deprecate dtype

* Release note + test fix

* Make deprecation of dtype argument pending
