+++ anndata/_io/specs/methods.py
@@ -246,8 +246,6 @@ def read_anndata(elem):
     ]:
         if k in elem:
             d[k] = read_elem(elem[k])
-        if "X" in d:
-            d["dtype"] = d["X"].dtype
     return AnnData(**d)
 
 
