+++ docs/release-notes/0.9.0.rst
@@ -7,6 +7,10 @@
 * Expanded docstring more documentation for ``backed`` argument of :func:`anndata.read_h5ad` :pr:`812` :smaller:`jeskowagner`
 * Added support for dask arrays :pr:`813` :smaller:`syelman` :smaller:`rahulbshrestha`
 
+.. rubric:: Breaking changes
+
+* The `AnnData` `dtype`` argument no longer defaults to `float32` :pr:`854` :smaller:`ivirshup`
+
 .. rubric:: Bug fixes
 
 .. rubric:: Updates
