### 2022-11-22 ac38062
Remove old deprecated force_dense argument
