+++ anndata/_io/h5ad.py
@@ -36,21 +36,10 @@ def write_h5ad(
     filepath: Union[Path, str],
     adata: AnnData,
     *,
-    force_dense: bool = None,
     as_dense: Sequence[str] = (),
     dataset_kwargs: Mapping = MappingProxyType({}),
     **kwargs,
 ) -> None:
-    if force_dense is not None:
-        warn(
-            "The `force_dense` argument is deprecated. Use `as_dense` instead.",
-            FutureWarning,
-        )
-    if force_dense is True:
-        if adata.raw is not None:
-            as_dense = ("X", "raw/X")
-        else:
-            as_dense = ("X",)
     if isinstance(as_dense, str):
         as_dense = [as_dense]
     if "raw.X" in as_dense:
