### 2022-11-22 0dc3cf9
Remove old deprecated force_dense argument (#855)
* Remove old deprecated force_dense argument

* release note
