### 2022-11-22 2253eec
Simplify key based dispatch test
