+++ anndata/tests/test_io_dispatched.py
@@ -15,14 +15,10 @@ from anndata.tests.helpers import gen_adata, assert_equal
 
 def test_read_dispatched_w_regex():
     def read_only_axis_dfs(func, elem_name: str, elem, iospec):
-        if elem_name == "/":
-            # TODO: this case is only complicated because of AnnData.__init__ dtype arg
-            d = {}
-            for k, v in elem.items():
-                v_read = read_dispatched(v, read_only_axis_dfs)
-                if v_read is not None:
-                    d[k] = v_read
-            return ad.AnnData(**d)
+        if iospec.encoding_type == "anndata":
+            return ad.AnnData(
+                **{k: read_dispatched(v, read_only_axis_dfs) for k, v in elem.items()}
+            )
         elif re.match(r"^/((obs)|(var))?(/.*)?$", elem_name):
             return func(elem)
         else:
