### 2022-11-22 5368f7d
Start hdf5 compat, currently broken
