### 2022-11-23 e1e029e
Actually fixed hdf5plugin references.
