+++ anndata/_core/anndata.py
@@ -1907,11 +1907,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             Filename of data file. Defaults to backing file.
         compression
             For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.   
-            
+
             Alternative compression filters such as `zstd` can be passed
             from the :doc:`hdf5plugin <hdf5plugin:usage>` library. 
             Experimental.
-            
+
             Usage example::
 
                 import hdf5plugin
@@ -1930,11 +1930,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
         compression_opts
             For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
-            
+
             Alternative compression filters such as `zstd` can be configured
             using helpers from the :doc:`hdf5plugin <hdf5plugin:usage>` 
             library. Experimental. 
-            
+
             Usage example (setting `zstd` compression level to 5)::
 
                 import hdf5plugin
