### 2022-11-23 5b37a88
Formatting changes to conform to trailing whitespace standards
