### 2022-11-23 db23929
Added documentation for hdf5plugin compression usage in write_h5ad
