### 2022-11-23 e675901
Backport PR #856: Fix order type (#858)
Co-authored-by: Isaac Virshup <ivirshup@gmail.com>
