### 2022-11-25 80b634b
(feat): r`ead_dispatched` with `zarr`
