### 2022-11-25 808dfbe
(feat): `read_dispatched` for h5ad
