### 2022-11-25 baae968
(feat): `write_dispatched` for `zarr`
