### 2022-11-28 7ce8191
(fix): remove unnecessary condition.
