### 2022-11-29 b99d0d4
move index types to .compat
