+++ anndata/tests/test_concatenate.py
@@ -1,8 +1,8 @@
-from collections.abc import Hashable, List, Callable
+from collections.abc import Hashable
 from copy import deepcopy
 from itertools import chain, product
 from functools import partial, singledispatch
-from typing import Any
+from typing import Any, List, Callable
 import warnings
 
 import numpy as np
