### 2022-11-29 57a602d
Fix flake8 config and updates flake8 to 6.0 (#863)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
