### 2022-11-29 1bf0020
Centralize pytest options, add -v
