### 2022-12-02 2f8dadf
Fix syntax highlighting in notebooks
