### 2022-12-02 6052c0c
Remove github links custom extension
