+++ docs/conf.py
@@ -52,15 +52,16 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.autosummary",
     "sphinx_autodoc_typehints",  # needs to be after napoleon
+    "sphinx_issues",
     "scanpydoc",
     "nbsphinx",
     "IPython.sphinxext.ipython_console_highlighting",
-    *[p.stem for p in (HERE / "extensions").glob("*.py")],
 ]
 
 # Generate the API documentation when building
 autosummary_generate = True
 autodoc_member_order = "bysource"
+issues_github_path = "scverse/anndata"
 # autodoc_default_flags = ['members']
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
