+++ docs/tutorials/index.rst
@@ -3,16 +3,12 @@ Tutorials
 
 .. TODO Figure out what text I want here, or at least fix links to tutorials
 
-For a quick introduction to `AnnData`, check out :tutorial:`getting-started`.
+For a quick introduction to `AnnData`, check out :doc:`Getting Started with AnnData <notebooks/getting-started>`.
 
 For working with the experimental data loaders (see :ref:`experimental_api`):
 
-* For an example of working with pytorch :tutorial:`annloader`
-* For lazily concatenating multiple files :tutorial:`anncollection`
-* For interfacing pytorch models with lazy collections :tutorial:`anncollection-annloader`
-
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 1
 
    notebooks/getting-started
    notebooks/annloader
