### 2022-12-02 3df9d6f
Fix up text on tutorials landing page
