+++ docs/tutorials/index.rst
@@ -1,11 +1,9 @@
 Tutorials
 =========
 
-.. TODO Figure out what text I want here, or at least fix links to tutorials
-
 For a quick introduction to `AnnData`, check out :doc:`Getting Started with AnnData <notebooks/getting-started>`.
 
-For working with the experimental data loaders (see :ref:`experimental_api`):
+For working with the experimental data loaders also see :ref:`experimental_api`.
 
 .. toctree::
    :maxdepth: 1
