### 2022-12-02 385bb5a
Add git submodule for tutorials (#852)
* turn off warn as error so we can look at docs while changing

* add submodule for notebooks

* Add submodule commands to readthedocs

* Change theme?

* Revert "Change theme?"

This reverts commit ec87af2fb9584eec2a361b755ce9057e89119066.

* Fix syntax highlighting in notebooks

* Capitalization?

* Add extension, ignore other files from tutorial repo

* Remove github links custom extension

* Turn warningiserror back on

* Fix up text on tutorials landing page
