+++ docs/release-notes/0.9.0.rst
@@ -20,4 +20,4 @@
 .. rubric:: Deprecations
 
 * :meth:`AnnData.concatenate() <anndata.AnnData.concatenate>` is now deprecated in favour of :func:`anndata.concat` :pr:`845` :smaller:`ivirshup`
-* Previously deprecated `force_dense` arugment :meth:`AnnData.write_h5ad() <anndata.AnnData.write_h5ad>` has been removed.:pr:`855` :smaller:`ivirshup`
+* Previously deprecated `force_dense` arugment :meth:`AnnData.write_h5ad() <anndata.AnnData.write_h5ad>` has been removed. :pr:`855` :smaller:`ivirshup`
