### 2022-12-02 04113eb
Remove deprecated slicing of adjacency matrices in uns
