+++ anndata/tests/test_base.py
@@ -323,21 +323,6 @@ def test_slicing_strings():
         adata[["A", "B", "not_in_obs"], :]
 
 
-def test_slicing_graphs():
-    # Testing for deprecated behaviour of connectivity matrices in .uns["neighbors"]
-    with pytest.warns(FutureWarning, match=r".obsp\['connectivities'\]"):
-        adata = AnnData(
-            np.array([[1, 2], [3, 4], [5, 6]]),
-            uns=dict(neighbors=dict(connectivities=sp.csr_matrix(np.ones((3, 3))))),
-        )
-
-    adata_sub = adata[[0, 1], :]
-    with pytest.warns(FutureWarning):
-        assert adata_sub.uns["neighbors"]["connectivities"].shape[0] == 2
-        assert adata.uns["neighbors"]["connectivities"].shape[0] == 3
-        assert adata_sub.copy().uns["neighbors"]["connectivities"].shape[0] == 2
-
-
 def test_slicing_series():
     adata = AnnData(
         np.array([[1, 2], [3, 4], [5, 6]]),
