### 2022-12-05 13d9d64
Fix tests for simplified error message
