### 2022-12-05 9e1078e
Fix sklearn package name
Signed-off-by: zethson <lukas.heumos@posteo.net>
