### 2022-12-05 1545b7b
Check for key rather than KeyError
KeyError's could come up for a variety of reasons during IO, we only want to throw a missing spec error when we don't have a spec.
