### 2022-12-05 259a9b7
Consolidate registry errors to one type
