+++ anndata/tests/test_io_elementwise.py
@@ -15,7 +15,6 @@ import zarr
 from anndata._io.utils import AnnDataReadError, AnnDataWriteError
 from anndata.compat import _read_attr, H5Group, ZarrGroup
 from anndata._io.specs import write_elem, read_elem
-from anndata._io.specs.registry import NoSuchIO
 from anndata.tests.helpers import assert_equal, gen_adata
 
 
