### 2022-12-05 ab2e4af
Revert addition of AnnDataWriteError
