+++ anndata/compat/__init__.py
@@ -43,19 +43,7 @@ except ImportError:
 
 
 try:
-    try:
-        from importlib.metadata import version
-    except ImportError:
-        from importlib_metadata import version
-
-    # Ensure compatibility with both 1.9rc and 2.x releases of awkward
-    # TODO Once 2.x is released, we can require it as a minimal dependency and remove this code.
-    awkward_version, _ = version("awkward").split(".", maxsplit=1)
-    awkward_version = int(awkward_version)
-    if awkward_version < 2:
-        import awkward._v2 as awkward
-    else:
-        import awkward
+    import awkward
 
     AwkArray = awkward.Array
 
