### 2023-01-02 fe27b74
Exclude awkward arrays from fill_value concat test
