+++ docs/fileformat-prose.md
@@ -1,111 +1,117 @@
-On-disk format
---------------
+# On-disk format
 
-.. note::
-   These docs are written for anndata 0.8.
-   Files written before this version may differ in some conventions,
-   but will still be read by newer versions of the library.
+```{note}
+These docs are written for anndata 0.8.
+Files written before this version may differ in some conventions,
+but will still be read by newer versions of the library.
+```
 
-AnnData objects are saved on disk to hierarchical array stores like HDF5_
-(via :doc:`H5py <h5py:index>`) and :doc:`zarr:index`.
+AnnData objects are saved on disk to hierarchical array stores like [HDF5]
+(via {doc}`H5py <h5py:index>`) and {doc}`zarr:index`.
 This allows us to have very similar structures in disk and on memory.
 
 As an example we’ll look into a typical `.h5ad` object that’s been through an analysis.
 This structure should be largely equivalent to Zarr structure, though there are a few minor differences.
 
-.. _HDF5: https://en.wikipedia.org/wiki/Hierarchical_Data_Format
-.. I’ve started using h5py since I couldn’t figure out a nice way to print attributes from bash.
+% I’ve started using h5py since I couldn’t figure out a nice way to print attributes from bash.
 
+```python
 >>> import h5py
 >>> f = h5py.File("02_processed.h5ad", "r")
 >>> list(f.keys())
 ['X', 'layers', 'obs', 'obsm', 'uns', 'var', 'varm']
+```
 
-.. .. code:: bash
-
-..    $ h5ls 02_processed.h5ad
-..    X                        Group
-..    layers                   Group
-..    obs                      Group
-..    obsm                     Group
-..    uns                      Group
-..    var                      Group
-..    varm                     Group
+<!-- ```bash
+$ h5ls 02_processed.h5ad
+X                        Group
+layers                   Group
+obs                      Group
+obsm                     Group
+uns                      Group
+var                      Group
+varm                     Group
+``` -->
 
 In general, `AnnData` objects are comprised of a various types of elements.
 Each element is encoded as either an Array (or Dataset in hdf5 terminology) or a collection of elements (e.g. Group) in the store.
 We record the type of an element using the `encoding-type` and `encoding-version` keys in it's attributes.
 For example, we can this file represents an `AnnData` object from this metadata:
 
+```python
 >>> dict(f.attrs)
 {'encoding-type': 'anndata', 'encoding-version': '0.1.0'}
+```
 
 Using this information, we're able to dispatch onto readers for the different element types that you'd find in an anndata.
 
-Dense arrays
-~~~~~~~~~~~~
+## Dense arrays
 
 Dense numeric arrays have the most simple representation on disk,
-as they have native equivalents in H5py :doc:`h5py:high/dataset` and Zarr :ref:`Arrays <zarr:tutorial_create>`.
+as they have native equivalents in H5py {doc}`h5py:high/dataset` and Zarr {ref}`Arrays <zarr:tutorial_create>`.
 We can see an example of this with dimensionality reductions stored in the `obsm` group:
 
+```python
 >>> f["obsm"].visititems(print)
 X_pca <HDF5 dataset "X_pca": shape (38410, 50), type "<f4">
 X_umap <HDF5 dataset "X_umap": shape (38410, 2), type "<f4">
 
 >>> dict(f["obsm"]["X_pca"].attrs)
[…]
