+++ pyproject.toml
@@ -72,6 +72,7 @@ doc = [
     "scanpydoc>=0.7.7",
     "zarr",
     "IPython", # For syntax highlighting in notebooks
+    "myst_parser",
 ]
 test = [
     "loompy>=3.0.5",
