### 2023-01-23 c57dcd0
add string + categorical encodings
