### 2023-01-23 cc465b9
Add scalar specifications
