+++ docs/conf.py
@@ -31,6 +31,7 @@ release = version
 # default settings
 templates_path = ["_templates"]
 html_static_path = ["_static"]
+source_suffix = [".rst", ".md"]
 master_doc = "index"
 default_role = "literal"
 exclude_patterns = [
