### 2023-01-23 3e2dda8
Improve interpretability of IOSpec errors (#734)
* Improve typing of IO registry

* Improved error message and some basic tests

* style

* split read and write errors

* Raise AnnDataWriteError

* Fix tests

* typo

* Fix doc links

* Add read test

* Add update help

* Add write tests

* Fix tests for simplified error message

* Try fixing doc builds, remove unused deps

* Check for key rather than KeyError

KeyError's could come up for a variety of reasons during IO, we only want to throw a missing spec error when we don't have a spec.

* Consolidate registry errors to one type

* Revert ad
[…]
