### 2023-01-23 0ba56a6
move 10x reader tests
