+++ anndata/_io/read.py
@@ -1,7 +1,9 @@
+from __future__ import annotations
+
 from pathlib import Path
 from os import PathLike, fspath
 from types import MappingProxyType
-from typing import Union, Optional, Mapping, Tuple
+from typing import Union, Optional, Literal, Mapping, Tuple
 from typing import Iterable, Iterator, Generator
 from collections import OrderedDict
 import gzip
@@ -491,3 +493,280 @@ def del_sparse_matrix_keys(mapping, key_csr):
     del mapping[f"{key_csr}_indices"]
     del mapping[f"{key_csr}_indptr"]
     del mapping[f"{key_csr}_shape"]
+
+
+# Reading 10x formats
+
+
+def read_10x_h5(
+    filename: Union[str, Path],
+    *,
+    genome: Optional[str] = None,
+    feature_types: str | list[str] | None = None,
+) -> AnnData:
+    """\
+    Read 10x-Genomics-formatted hdf5 file.
+
+    Parameters
+    ----------
+    filename
+        Path to a 10x hdf5 file.
+    genome
+        Filter expression to genes within this genome. For legacy 10x h5
+        files, this must be provided if the data contains more than one genome.
+    Feature types
+        Which feature types to read in. Reads all in by default.
+
+    Returns
+    -------
+    Annotated data matrix, where observations/cells are named by their
+    barcode and variables/genes by gene name. Stores the following information:
+
+    :attr:`~anndata.AnnData.X`
+        The data matrix is stored
+    :attr:`~anndata.AnnData.obs_names`
+        Cell names
+    :attr:`~anndata.AnnData.var_names`
+        Gene names
+    :attr:`~anndata.AnnData.var`\\ `['gene_ids']`
+        Gene IDs
+    :attr:`~anndata.AnnData.var`\\ `['feature_types']`
+        Feature types
+    """
+    # Coerce feature_type to be either list of string or None
+    if feature_types is not None:
+        if isinstance(feature_types, str):
+            feature_types = [feature_types]
+
+    with h5py.File(str(filename), "r") as f:
+        v3 = "/matrix" in f
+    if v3:
+        adata = _read_v3_10x_h5(filename)
+        if genome:
+            if genome not in adata.var["genome"].values:
+                raise ValueError(
+                    f"Could not find data corresponding to genome '{genome}' in '{filename}'. "
+                    f'Available genomes are: {list(adata.var["genome"].unique())}.'
+                )
+            adata = adata[:, adata.var["genome"] == genome]
+        if feature_types is not None:
+            adata = adata[:, adata.var["feature_types"].isin(feature_types)]
+        if adata.is_view:
+            adata = adata.copy()
+    else:
+        adata = _read_legacy_10x_h5(filename, genome=genome)
+    return adata
+
+
+def _read_v3_10x_h5(filename):
+    """
+    Read hdf5 file from Cell Ranger v3 or later versions.
+    """
+    with h5py.File(str(filename), "r") as f:
+        dsets = {}
+        _collect_datasets(dsets, f["matrix"])
+
+        from scipy.sparse import csr_matrix
+
+        M, N = dsets["shape"]
+        data = dsets["data"]
+        if dsets["data"].dtype == np.dtype("int32"):
+            data = dsets["data"].view("float32")
+            data[:] = dsets["data"]
+        matrix = csr_matrix(
+            (data, dsets["indices"], dsets["indptr"]),
+            shape=(N, M),
+        )
+        adata = AnnData(
+            matrix,
+            obs=dict(obs_names=dsets["barcodes"].astype(str)),
+            var=dict(
+                var_names=dsets["name"].astype(str),
+                gene_ids=dsets["id"].astype(str),
[…]
