### 2023-01-23 f36501b
Apply more suggestions from code review
