### 2023-01-25 af54e94
Add not that obs, var dataframes don't need more than an index
