### 2023-01-25 7a12515
Revert changes to some backwards compat tests
