### 2023-01-25 9736765
Revert changes to some backwards compat tests
