+++ anndata/_io/utils.py
@@ -175,22 +175,46 @@ def report_read_key_on_error(func):
     >>> z["X"] = [1, 2, 3]
     >>> read_arr(z["X"])  # doctest: +SKIP
     """
+    from inspect import ismethod
 
-    @wraps(func)
-    def func_wrapper(elem, *args, **kwargs):
-        try:
-            return func(elem, *args, **kwargs)
-        except Exception as e:
-            if isinstance(e, AnnDataReadError):
-                raise e
-            else:
-                parent = _get_parent(elem)
-                raise AnnDataReadError(
-                    f"Above error raised while reading key {elem.name!r} of "
-                    f"type {type(elem)} from {parent}."
-                ) from e
+    def re_raise_error(e, elem):
+        if isinstance(e, AnnDataReadError):
+            raise e
+        else:
+            parent = _get_parent(elem)
+            raise AnnDataReadError(
+                f"Above error raised while reading key {elem.name!r} of "
+                f"type {type(elem)} from {parent}."
+            ) from e
 
-    return func_wrapper
+    if ismethod(func):
+
+        @wraps(func)
+        def method_wrapper(self, elem, *args, **kwargs):
+            try:
+                return func(self, elem, *args, **kwargs)
+            except Exception as e:
+                re_raise_error(e, elem)
+
+        return method_wrapper
+
+    else:
+
+        # TODO: sometimes, something that looks an awful lot like a method is reaching here
+        # It's being passed an instance of a Reader, has a signature Reader.read_elem, but is a function
+        @wraps(func)
+        def func_wrapper(*args, **kwargs):
+            from anndata._io.specs import Reader
+
+            for elem in args:
+                if not isinstance(elem, Reader):
+                    break
+            try:
+                return func(*args, **kwargs)
+            except Exception as e:
+                re_raise_error(e, elem)
+
+        return func_wrapper
 
 
 def report_write_key_on_error(func):
@@ -207,21 +231,37 @@ def report_write_key_on_error(func):
     >>> X = [1, 2, 3]
     >>> write_arr(z, "X", X)  # doctest: +SKIP
     """
+    from inspect import ismethod
 
-    @wraps(func)
-    def func_wrapper(elem, key, val, *args, **kwargs):
-        try:
-            return func(elem, key, val, *args, **kwargs)
-        except Exception as e:
-            if "Above error raised while writing key" in format(e):
-                raise
-            else:
-                parent = _get_parent(elem)
-                raise type(e)(
-                    f"{e}\n\n"
-                    f"Above error raised while writing key {key!r} of {type(elem)} "
-                    f"to {parent}"
-                ) from e
+    def re_raise_error(e, elem, key):
+        if "Above error raised while writing key" in format(e):
+            raise
+        else:
+            parent = _get_parent(elem)
+            raise type(e)(
+                f"{e}\n\n"
+                f"Above error raised while writing key {key!r} of {type(elem)} "
+                f"to {parent}"
+            ) from e
+
+    # Need to specialize for method signature
+    if ismethod(func):
+
+        @wraps(func)
+        def func_wrapper(self, elem, key, val, *args, **kwargs):
+            try:
+                return func(self, elem, key, val, *args, **kwargs)
+            except Exception as e:
+                re_raise_error(e, elem, key)
+
+    else:
+
+        @wraps(func)
+        def func_wrapper(elem, key, val, *args, **kwargs):
+            try:
+                return func(elem, key, val, *args, **kwargs)
+            except Exception as e:
+                re_raise_error(e, elem, key)
 
     return func_wrapper
 
