### 2023-01-25 49e8069
Start fixing error reporting
To fix error reporting, I've put the attempt to catch an error during IO on top of the `read_elem` method.
Since the decorator is sometimes used on functions, I modified it to be able to handle the signature of both a method and a function.

What's weird is that sometimes the decorator is being passed the arguments of a method, that has a name like a method, but is a function.
So that still needs to be fixed.
