### 2023-01-26 fccdde5
Update categorical spec to include index base and null values
