+++ benchmark/benchmarks/readwrite.py
@@ -118,7 +118,7 @@ class H5ADReadSuite:
         mem_recording = memory_usage(
             (sedate(anndata.read_h5ad, 0.005), (self.filepath,)), interval=0.001
         )
-        adata = anndata.read_h5ad(self.filepath)
+        # adata = anndata.read_h5ad(self.filepath)
         base_size = mem_recording[-1] - mem_recording[0]
         print(np.max(mem_recording) - np.min(mem_recording))
         print(base_size)
