+++ benchmark/download_dataset.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+mkdir benchmarks/data
+cd benchmarks/data
+wget http://cf.10xgenomics.com/samples/cell-exp/1.1.0/pbmc3k/pbmc3k_filtered_gene_bc_matrices.tar.gz -O pbmc3k_filtered_gene_bc_matrices.tar.gz
+tar -xzf pbmc3k_filtered_gene_bc_matrices.tar.gz
+python ../../fetch_datasets.py
