+++ docs/benchmarks.rst
@@ -9,4 +9,3 @@ Below follows a simple benchmark showing read-write efficiency.
    :maxdepth: 1
 
    benchmark-read-write
-   benchmark-comparitive
