+++ .github/workflows/benchmark.yml
@@ -2,9 +2,9 @@ name: check-benchmarks
 
 on:
     push:
-        branches: [main]
+        branches: [master]
     pull_request:
-        branches: [main]
+        branches: [master]
 
 jobs:
     check-benchmarks:
