### 2023-01-27 a589820
Add IO testcase for AIRR data
