### 2023-01-30 0ea1a70
Remove extra notebook
