+++ docs/index.rst
@@ -30,7 +30,6 @@ Latest additions
    concatenation
    fileformat-prose
    interoperability
-   anndata_dask_array
    benchmarks
    contributing
    release-notes/index
