### 2023-04-17 scverse/anndata#984 @falexwolf
Finalize anndata paper
Re: https://github.com/openjournals/joss-reviews/issues/4371#issuecomment-1511258256

@ivirshup, another attempt. This the thread in which JOSS editors are checking in. It'd be incredible if you responded.
