### 2023-04-29 scverse/anndata#769 @rcannood
I met up with @ivirshup during the [scverse hackathon](https://scverse.org/events/2023_04_hackathon/). He mentioned that there are benchmarks on ±100'000 cells performed by a CI. I'm eager to see the results :)
