### 2023-06-07 scverse/anndata#769 @rcannood
Having passed the 1 year anniversary for the AnnData JOSS submission, I'd really like to have this submission wrapped up and dealt with. AnnData is a great resource to researchers all over the world, and I think it'd be a pity if this work doesn't get published at JOSS.

@falexwolf The paper mentions:

> Furthermore, anndata is systematically benchmarked for performance using airspeed velocity [@Droettboom13], with the results linked from the docs.

@falexwolf Can you remind me where these results are published? The [Benchmarks](https://anndata.readthedocs.io/en/latest/benchmarks.html) page only links to a [simple benchmark](https://anndata.readthedocs.io/en/latest/benchmark-read-write.html) and the [ivirshup/anndata-benchmarks](https://github.com/ivirshup/anndata-benchmarks) repository, but I don't see any actual results. Could you help me find them?
