### 2023-06-07 scverse/anndata#769 @falexwolf
All benchmarking referenced here is by @ivirshup, I'm not in the loop on the details. @ivirshup, could you help out?
