### 2023-08-10 #4371 @falexwolf
It's very understandable that you'd retract the submission. I'm sorry for the effort we caused. At least, it seems the basic idea of having a paper for anndata wasn't wrong and the preprint is cited with 75 citations according to https://badge.dimensions.ai/details/id/pub.1144066001.

The blocker consists in a few technical points that only the first author (@ivirshup) can answer.

All other points have a response in the thread we have with the reviewers (https://github.com/scverse/anndata/issues/769). I'll happily move them here should @ivirshup make the response complete.
