### 2023-08-10 #4371 @rcannood
Though I think having the joint collaboration of the authors to tackle the remaining issues in scverse/anndata#769 would be the best outcome, I don't think it's the only solution.

In an attempt to get this submission finally accepted, let me provide a recap of the only remaining blocking item in my review at scverse/anndata#769:

<details><summary>Recap of blocking item related to "Performance"</summary>

* [JOSS] **Performance**: If there are any performance claims of the software, have they been confirmed?
* [@rcannood] There are various claims throughout the paper that anndata takes into account scRNAseq's dataset sizes in terms of cpu and memory usage. For example:
  -  "Due to the increasing scale of data, we emphasized efficient operations with low memory and runtime overhead."
  -  "In particular, AnnData takes great pains to support efficient operations with sparse data."
* [@rcannood] I totally agree with the authors on this matter. However, I don't see any benchmarks in the paper or the documentation where anndata is compared to SCE, Seurat and/or loom in terms of memory usage, disk usage and/or execution time.
* [@falexwolf] Thank you for pointing out that we didn’t substantiate performance claims in the paper! We’re glad to read that you agree with us on anndata’s good performance in comparison to many other data structures in the field.
* [@falexwolf] Within the docs, we have this page that serves as a landing page for benchmarks: https://anndata.readthedocs.io/en/latest/benchmarks.html. To date, it has two elements:
  - It makes a comparison with the loom on-disk format here: https://anndata.readthedocs.io/en/latest/benchmark-read-write.html
  - It links to https://github.com/ivirshup/anndata-benchmarks, an entire devoted to benchmarking anndata over its entire lifetime using [airspeed velocity](https://asv.readthedocs.io/en/stable/).
* [@falexwolf] We will add a more comprehensive comparative benchmarks in the docs, link it from the landing page and maintain it.
* [@rcannood] The read/write benchmark is a good start! This only covers a very specific aspect of AnnData and only compares to loom. I'm not sure I'm wholly satisfied that this is sufficient to substantiate your claims regarding performance. The manuscript states: "Due to the increasing scale of data, we emphasized efficient operations with low memory and runtime overhead."
    - "Increasing scale of data" → the dataset used in this benchmark only contains 2300 cells
    - "efficient operations" → the same paragraph lists these operations:
      * out of core conversions between dense and sparse data
      * lazy subsetting
      * per-element operations for low total memory usage
      * in-place subsetting
      * combining `AnnData` objects with various merge strategies
      * lazy concatenation
      * batching
      * backed out-of-memory mode
    - "with low memory [...] overhead" → the benchmark currently doesn't cover memory overhead.
    - "low [...] runtime overhead" → the current benchmark compares against loom to demonstrate runtime efficiency. I wonder, how does it stack up against SCE when performing similar operations?

IMO, I think these are the following options for resolving the matter:

* Option 1: Resolve the issue by publishing more extensive benchmarks to back up the claims made in the manuscript.
* Option 2: If that is not possible, the performance claims made in the paper could be removed or watered down, as the paper is still valuable even without making any claims related to performance.
* Option 3: If the performance claims in the manuscript are essential but extensive benchmarks will take too long to develop, perhaps the JOSS editors can make an editorial decision on whether such benchmarks are needed to back up the performance claims made in the manuscript.

Hopefully this helps in figuring out a solution to the remaining issues.
