### 2023-08-10 #4371 @falexwolf
Thank you, @rcannood - you're contributions are beyond anything that could be expected from any reviewer.

I'm happy to implement Option 2 if this is something reviewers and editors can accept. Option 3, I guess, equates to an acceptance of the current revision. Option 1 is what I can't do because it's highly technical and @ivirshup has been working on this for years (and is still working on it with the team, I believe).
