### 2023-09-08 #4371 @rcannood
Hey @falexwolf ! Sorry for not responding earlier, it wasn't clear to me that you were posing a question.

Options 1 to 3 are all viable solutions for me :) :+1:
