### 2023-10-04 #4371 @rcannood
@Kevin-Mattheus-Moerman @luizirber Can I just check the `[ ] Performance` checkbox and quietly get this over with?

While I think the benchmarks themselves would be super interesting; at this stage I feel like I'm doing the scientific community a disservice for not allowing the AnnData paper to be published, since the performance of AnnData is not really the main point of the package or the accompanying manuscript.
