### 2023-12-11 #4371 @luizirber
sorry for the unresponsiveness :disappointed:

> @Kevin-Mattheus-Moerman @luizirber Can I just check the `[ ] Performance` checkbox and quietly get this over with?
>
> While I think the benchmarks themselves would be super interesting; at this stage I feel like I'm doing the scientific community a disservice for not allowing the AnnData paper to be published, since the performance of AnnData is not really the main point of the package or the accompanying manuscript.

I agree. Not that performance is not important, but option 2 seems the best to proceed because the main focus is the usability and the use cases enabled by the library.
