### 2023-12-12 #4371 @rcannood
Alright, in that case I'll leave it in the hands of @falexwolf and colleagues to make changes to the paper.

@luizirber Could you run `@editorialbot set https://github.com/scverse/anndata as repository` ?
