### 2023-12-15 #4371 @falexwolf
Performance claims are downtuned: https://github.com/scverse/anndata/pull/1267

With this, all reviewer concerns should be addressed, if I'm not mistaken.
