### 2023-12-15 scverse/anndata#769 @falexwolf
Downtuned performance claims as requested by the editor: https://github.com/scverse/anndata/pull/1267

See discussion in the JOSS review: https://github.com/openjournals/joss-reviews/issues/4371
