### 2023-12-15 scverse/anndata#769 @falexwolf
This completes a sequence of 3 PRs to the paper branch
- https://github.com/scverse/anndata/pull/779
- https://github.com/scverse/anndata/pull/825
- https://github.com/scverse/anndata/pull/1267

For reference, the paper branch is here:
- https://github.com/scverse/anndata/pull/666
