### 2023-12-16 scverse/anndata#769 @rcannood
Thanks @falexwolf! This resolves all of my outstanding comments :) Closing this issue.
