### 2023-12-16 #4371 @rcannood
@luizirber Having thoroughly reviewed the manuscript and completed the last item in the evaluation checklist, I am pleased to inform you of my strong recommendation for its publication. The manuscript is exceptionally well-written and to the point.  Furthermore, anndata is already the _de facto_ standard file format for performing single cell analysis in Python.

In light of these considerations, I firmly believe that the manuscript not only meets but exceeds the publication standards of JOSS.

Thank you for the opportunity to review this work. I look forward to seeing the manuscript published in JOSS :)
