### 2024-02-12 #4371 @nlhepler
@luizirber I am pleased to say that having reviewed the latest revision of the paper I can recommend its publication. I concur with @rcannood that the manuscript is well-written, and that the software and format speak for themselves.

Looking forward to seeing it published in JOSS. :) Thanks for the opportunity, it's been a pleasure.
