### 2024-02-14 #4371 @Kevin-Mattheus-Moerman
@luizirber @falexwolf may I suggest/recommend that the paper title is extended/altered. E.g. using the format `ProjectName: Sentence saying what the project does`, the current title is rather short and vague.
