### 2024-05-25 #4371 @Kevin-Mattheus-Moerman
@openjournals/dev the check references command has not been working here, can  you check what is going on?
