### 2024-05-26 #4371 @xuanxu
The metadata of the paper file is [referencing the bib file](https://github.com/scverse/anndata/blob/paper/paper/paper.md?plain=1#L34) incorrectly. That should read:
```
bibliography: paper.bib
```
