### 2024-05-27 2d04dd9
Fix bibliography reference
