### 2024-05-27 #4371 @falexwolf
Done! https://github.com/scverse/anndata/commit/2d04dd9c4097cda847fd3853e1af0f8e9324d7db
