### 2024-07-18 #4371 @falexwolf
Done:

- https://github.com/scverse/anndata/pull/1566
