+++ paper/paper.bib
@@ -101,8 +101,7 @@ keywords = {}}
 @Proceedings{McKinney2010,
 author = {McKinney, Wes},
 title = {Data structures for statistical computing in python},
-booktitle = {Data structures for statistical computing in python},
-volume = {Proceedings of the 9th Python in Science Conference 445},
+booktitle = {Proceedings of the 9th Python in Science Conference 445},
 publisher = {Austin, TX},
 doi = {10.25080/majora-92bf1922-00a},
 pages = {51-56},
@@ -175,6 +174,7 @@ keywords = {}}
 author = {Ramos, Marcel and Schiffer, Lucas and Re, Angela and Azhar, Rimsha and Basunia, Azfar and Rodriguez, Carmen and Chan, Tiffany and Chapman, Phil and Davis, Sean R. and Gomez-Cabrero, David and Culhane, Aedin C. and Haibe-Kains, Benjamin and Hansen, Kasper D. and Kodali, Hanish and Louis, Marie S. and Mer, Arvind S. and Riester, Markus and Morgan, Martin and Carey, Vince and Waldron, Levi},
 title = {Software for the Integration of Multiomics Experiments in Bioconductor},
 journal = {Cancer Research},
+doi = {10.1158/0008-5472.can-17-0344},
 volume = {77},
 number = {21},
 pages = {e39--e42},
