### 2024-07-18 #4371 @falexwolf
Is fixed: https://github.com/scverse/anndata/pull/666/commits/a5166c56cf9794039663482f425f0541913ca69f
