### 2024-07-21 #4371 @Kevin-Mattheus-Moerman
- [ ] Please let us know what version tag is most appropriate to use. I've just set the version as `0.10.8`. As I see this is your latest tagged release.
- [ ] If it is `0.10.8` please create a new/updated ZENODO archive for this version (the one listed above is for February) and let me know what the ZENODO DOI is. For the ZENODO archive, please ensure that:
* The title matches the paper title
* The author set matches the paper author set
* That the license  matches your software license (Note for instance that the license listed for the current one is incorrect too: https://zenodo.org/records/10705762).

On the paper:
- [ ] Please check the above reference check. Can you verify that you've added all possible DOI's, and that you think there are no DOI's for those remaining references?
- [ ] The paper has been reviewed already, and I've re-checked it as well for typos etc. However, it is rather lengthy. Although its current form is permissible, I would encourage you to review the paper's "non-JOSS-required" sections, and check they can be shortened/removed e.g. if the information is already present in the documentation. If you prefer to keep it as is just let me know.
