### 2024-08-28 #4371 @rcannood
@Kevin-Mattheus-Moerman GJ on sending frequent reminders :+1:

I wonder whether @falexwolf feels partially overwhelmed by this item:

> However, it is rather lengthy. Although its current form is permissible, I would encourage you to review the paper's "non-JOSS-required" sections, and check they can be shortened/removed e.g. if the information is already present in the documentation. If you prefer to keep it as is just let me know.

Reducing the word count of a manuscript is typically quite time intensive. However, you did offer Alex a way out of this so there's that ... :relaxed:
