### 2024-09-03 #4371 @falexwolf
Apologies again for the radio silence.

To your remaining requirements:

- Zenodo & version: I updated the zenodo record and the version: `0.10.9` /  https://zenodo.org/records/13643180
- DOIs: Yes, I went through several iterations of the reference check a few weeks ago to add all DOIs (see logging output above). The remaining references do not have DOIs.
- Paper length: I prefer to keep it as is. Thank you for offering this way out.
