### 2024-09-05 #4371 @Kevin-Mattheus-Moerman
@openjournals/dev can you help with this paper compilation error? It seems to be related to affiliations but I can't quite make out what is wrong. Thanks
