### 2024-09-05 #4371 @danielskatz
This isn't the issue with compiling, but looking at the affiliations, they should all have the final period "." removed, and have the country added. The period at the end of the note for Wolf should also be removed.
