### 2024-09-05 #4371 @falexwolf
> Can you please edit the archive listed license to be 0.10.9, it currently says v2. There is no need to create a new archive link any more at this point, you can manually edit the listing and version tag.

Done!

> This isn't the issue with compiling, but looking at the affiliations, they should all have the final period "." removed, and have the country added. The period at the end of the note for Wolf should also be removed.

I'll try this right away.
