### 2024-09-05 #4371 @falexwolf
Removed the periods and added the countries: https://github.com/scverse/anndata/pull/666/commits/4ca388ae5ac4f4bbf49c23fe9bdd46aaa3ce7278
