### 2024-09-08 #4371 @rcannood
@Kevin-Mattheus-Moerman Does the proof look good to you?
