### 2024-09-10 #4371 @Kevin-Mattheus-Moerman
@rcannood thanks for your help with this review. I'll pick this up shortly.

I was wondering, can you please re-confirm if you have anything that an outsider might perceive as a conflict of interest with this submission? You are not listed as an author, but do you actively work with the authors? I see you are a contributor to the project, was that during this review or before? Thanks.
