### 2024-09-10 #4371 @rcannood
@Kevin-Mattheus-Moerman I'd say back in May 2022 when I first reviewed the publication, I did not have any COIs related to the manuscript. Back then, I had probably already talked to some of the manuscript authors via GitHub issues, but did not know any of the authors personally.

Since then, I have become an avid user of the package, and use it in a lot of my projects. I did not contribute to the conceptualisation, development, or maintenance of this package, other than my review and one minor contribution I made in Dec 2020 (See below).

My enthusiasm related to wanting this paper published comes from my general affinity towards OSS and by extension JOSS, and also my belief that anndata is a useful package and its publication deserves publication.

However, for full transparency, I did start development on an R package called [laminlabs/laminr](https://github.com/laminlabs/laminr) last month in for Lamin Labs, whose CEO is @falexwolf -- the submitter of this manuscript.

I hope this addresses any concerns you may have.

TL;DR: I did not at the time of the review, I suppose that I do now.

------

### Additional information

In Dec 2020 I made a small open-source contribution (https://github.com/scverse/anndata/pull/463) related to an issue that I was encountering (https://github.com/scverse/anndata/issues/464).

This issue I encountered while creating an R package also called 'anndata' ([CRAN](https://cran.r-project.org/web/packages/anndata/index.html)) which exposes the Python interface to  users via reticulate. This was developed fully by myself without input from the scverse organisation  ([dynverse/anndata](https://github.com/dynverse/anndata/graphs/contributors)).

I did end up creating [scverse/anndataR](github.com/scverse/anndataR) in collaboration with other non-scverse members lazappi LouiseDck and mtmorgan. However, this work was started in April 2023 as part of the scverse hackathon in Heidelberg (https://scverse.org/events/2023_04_hackathon/), which is way past my initial review.
